Move Models headers to include directory
