+++ include/models/MultipleClassifiers.hh
@@ -6,11 +6,11 @@
 #ifndef _MULTCLASS_HH_
 #define _MULTCLASS_HH_
 
-#include "../Models/C45Tree.hh"
-#include "../Models/M5Tree.hh"
-#include "../Models/LinearSplitsTree.hh"
+#include "models/C45Tree.hh"
+#include "models/M5Tree.hh"
+#include "models/LinearSplitsTree.hh"
 
-#include "../Models/Stump.hh"
+#include "models/Stump.hh"
 
 #include <rl_common/Random.h>
 #include <rl_common/core.hh>
