+++ src/models/LinearSplitsTree.cc
@@ -1,123 +1,116 @@
-/** \file M5Tree.cc
-    Implements the M5 Decision tree, as described in:
-    "Learning with Continuous Classes" by J.R. Quinlan
-    "Inducing Model Trees for Continuous Classes" by Y. Wang and I.H. Witten
-    \author Todd Hester
-*/
+#include "models/LinearSplitsTree.hh"
 
-#include "M5Tree.hh"
-
+// LinearSplitsTree, from the following sources:
 
 // Include stuff for newmat matrix libraries
 
-#define WANT_MATH                // include.h will get math fns
-                                 // newmatap.h will get include.h
-#include <newmat/newmatap.h>     // need matrix applications
+#define WANT_MATH                    // include.h will get math fns
+                                     // newmatap.h will get include.h
+#include <newmat/newmatap.h>         // need matrix applications
 #ifdef use_namespace
-using namespace NEWMAT;             // access NEWMAT namespace
+using namespace NEWMAT;              // access NEWMAT namespace
 #endif
 
+// TODO:
+//  - save regression from split testing rather than re-building it later
 
 
-M5Tree::M5Tree(int id, int trainMode, int trainFreq, int m,
-               float featPct, bool simple, bool allowAllFeats, 
-	       float min_sdr, Random rng):
-  id(id), mode(trainMode), freq(trainFreq), M(m),
-  featPct(featPct), SIMPLE(simple), ALLOW_ALL_FEATS(allowAllFeats),
-  MIN_SDR(min_sdr), rng(rng)
+LinearSplitsTree::LinearSplitsTree(int id, int trainMode, int trainFreq, int m,
+                                   float featPct, bool simple, float min_er,
+                                   Random rng):
+  id(id), mode(trainMode), 
+  freq(trainFreq), M(m),
+  featPct(featPct), SIMPLE(simple), 
+  MIN_ER(min_er), rng(rng)
 {
 
   nnodes = 0;
   nOutput = 0;
   nExperiences = 0;
   hadError = false;
+  maxnodes = N_LS_NODES;
   totalnodes = 0;
-  maxnodes = N_M5_NODES;
-
 
   // how close a split has to be to be randomly selected
   SPLIT_MARGIN = 0.0; //0.02; //5; //01; //0.05; //0.2; //0.05;
 
-  LMDEBUG = false;
-  DTDEBUG = false;///true;
-  SPLITDEBUG = false;//true;
+  LMDEBUG = false;// true;
+  DTDEBUG = false;//true;
+  SPLITDEBUG = false; //true;
   STOCH_DEBUG = false; //true; //false; //true;
   INCDEBUG = false; //true; //false; //true;
   NODEDEBUG = false;
   COPYDEBUG = false; //true;
-  nfeat = 4;
 
-  cout << "Created m5 decision tree " << id;
+  cout << "Created linear splits decision tree " << id;
   if (SIMPLE) cout << " simple regression";
-  else cout << " multivariate regression";
-  if (ALLOW_ALL_FEATS) cout << " (all feats)";
-  else cout << " (subtree feats)";
+  else cout << " multivariate regrssion";
   if (DTDEBUG){
     cout << " mode: " << mode << " freq: " << freq << endl;
   }
-  cout << " MIN_SDR: " << MIN_SDR << endl;
+  cout << " MIN_ER: " << MIN_ER << endl;
+
 
   initNodes();
   initTree();
 
-
 }
 
-M5Tree::M5Tree(const M5Tree& m5):
-  id(m5.id), mode(m5.mode), freq(m5.freq), M(m5.M),
-  featPct(m5.featPct), SIMPLE(m5.SIMPLE), ALLOW_ALL_FEATS(m5.ALLOW_ALL_FEATS),
-  MIN_SDR(m5.MIN_SDR), rng(m5.rng)
+LinearSplitsTree::LinearSplitsTree(const LinearSplitsTree& ls):
+  id(ls.id), mode(ls.mode), 
+  freq(ls.freq), M(ls.M),
+  featPct(ls.featPct), SIMPLE(ls.SIMPLE), 
+  MIN_ER(ls.MIN_ER), rng(ls.rng)
 {
-  COPYDEBUG = m5.COPYDEBUG;
-  if (COPYDEBUG) cout << "m5 copy " << id << endl;
+  COPYDEBUG = ls.COPYDEBUG;
+  if (COPYDEBUG) cout << "LS copy " << id << endl;
   nnodes = 0;
-  nOutput = m5.nOutput;
-  nExperiences = m5.nExperiences;
-  hadError = m5.hadError;
+  nOutput = ls.nOutput;
+  nExperiences = ls.nExperiences;
+  hadError = ls.hadError;
   totalnodes = 0;
-  maxnodes = m5.maxnodes;
-  SPLIT_MARGIN = m5.SPLIT_MARGIN; 
-  LMDEBUG = m5.LMDEBUG;
-  DTDEBUG = m5.DTDEBUG;
-  SPLITDEBUG = m5.SPLITDEBUG;
-  STOCH_DEBUG = m5.STOCH_DEBUG; 
-  INCDEBUG = m5.INCDEBUG; 
-  NODEDEBUG = m5.NODEDEBUG;
-  nfeat = m5.nfeat;
+  maxnodes = ls.maxnodes;
+  SPLIT_MARGIN = ls.SPLIT_MARGIN; 
+  LMDEBUG = ls.LMDEBUG;
+  DTDEBUG = ls.DTDEBUG;
+  SPLITDEBUG = ls.SPLITDEBUG;
+  STOCH_DEBUG = ls.STOCH_DEBUG; 
+  INCDEBUG = ls.INCDEBUG; 
+  NODEDEBUG = ls.NODEDEBUG;
 
-  if (COPYDEBUG) cout << "   M5 copy nodes, experiences, root, etc" << endl;
+  if (COPYDEBUG) cout << "   LS copy nodes, experiences, root, etc" << endl;
   // copy all experiences
-  for (int i = 0; i < N_M5_EXP; i++){
-    allExp[i] = m5.allExp[i];
+  for (int i = 0; i < N_LST_EXP; i++){
+    allExp[i] = ls.allExp[i];
   }
-  if (COPYDEBUG) cout << "   M5 copied exp array" << endl;
+  if (COPYDEBUG) cout << "   LS copied exp array" << endl;
 
   // set experience pointers
-  experiences.resize(m5.experiences.size());
-  for (unsigned i = 0; i < m5.experiences.size(); i++){
+  experiences.resize(ls.experiences.size());
+  for (unsigned i = 0; i < ls.experiences.size(); i++){
     experiences[i] = &(allExp[i]);
   }
-  if (COPYDEBUG) cout << "   M5 set experience pointers" << endl;
+  if (COPYDEBUG) cout << "   LS set experience pointers" << endl;
 
   // now the tricky part, set the pointers inside the tree nodes correctly
   initNodes();
 
-  if (COPYDEBUG) cout << "   M5 copy tree " << endl;
+  if (COPYDEBUG) cout << "   LS copy tree " << endl;
   root = allocateNode();
   lastNode = root;
-  copyTree(root, m5.root);
-  if (COPYDEBUG) cout << "   M5 tree copy done" << endl;
+  copyTree(root, ls.root);
+  if (COPYDEBUG) cout << "   LS  tree copy done" << endl;
    
   if (COPYDEBUG) {
     cout << endl << "New tree: " << endl;
     printTree(root, 0);
     cout << endl;
-    cout << "  m5 copy done" << endl;
+    cout << "  LS copy done" << endl;
   }
 
 }
 
-void M5Tree::copyTree(tree_node* newNode, tree_node* origNode){
+void LinearSplitsTree::copyTree(tree_node* newNode, tree_node* origNode){
 
   int nodeId = newNode->id;
 
@@ -148,27 +141,23 @@ void M5Tree::copyTree(tree_node* newNode, tree_node* origNode){
   }
 }
 
-M5Tree* M5Tree::getCopy(){
-  M5Tree* copy = new M5Tree(*this);
+LinearSplitsTree* LinearSplitsTree::getCopy(){
+  LinearSplitsTree* copy = new LinearSplitsTree(*this);
   return copy;
 }
 
-M5Tree::~M5Tree() {
+LinearSplitsTree::~LinearSplitsTree() {
   deleteTree(root);
-  for (unsigned i = N_M5_EXP; i < experiences.size(); i++){
+  for (unsigned i = N_LST_EXP; i < experiences.size(); i++){
     delete experiences[i];
   }
   experiences.clear();
 }
 
 // here the target output will be a single value
-bool M5Tree::trainInstance(classPair &instance){
+bool LinearSplitsTree::trainInstance(classPair &instance){
 
   if (DTDEBUG) cout << id << " trainInstance" << endl;
-
-  //  featPct *= 0.99;
-
-  nfeat = instance.in.size();
 
   bool modelChanged = false;
 
@@ -176,7 +165,7 @@ bool M5Tree::trainInstance(classPair &instance){
 
   // take from static array until we run out
   tree_experience *e;
-  if (nExperiences < N_M5_EXP){
+  if (nExperiences < N_LST_EXP){
     // from statically created set of experiences
     e = &(allExp[nExperiences]);
 
@@ -220,8 +209,6 @@ bool M5Tree::trainInstance(classPair &instance){
     cout << endl << " Now have " << nExperiences << " experiences." << endl;
   }
 
-  // depending on mode/etc, maybe re-build tree
-
   // mode 0: re-build every step
   if (mode == BUILD_EVERY || nExperiences <= 1){
     rebuildTree();
@@ -241,7 +228,7 @@ bool M5Tree::trainInstance(classPair &instance){
     if (error > 0.0){
       rebuildTree();
       modelChanged = true;
-    }
+    } 
   }
 
   // mode 2: re-build every FREQ steps
@@ -254,14 +241,35 @@ bool M5Tree::trainInstance(classPair &instance){
   }
 
   if (modelChanged){
-    if (DTDEBUG) cout << "DT " << id << " tree re-built." << endl;
+    if (DTDEBUG || SPLITDEBUG) cout << "DT " << id << " tree re-built." << endl;
 
-    if (DTDEBUG){
+    if (DTDEBUG || SPLITDEBUG){
       cout << endl << "DT: " << id << endl;
       printTree(root, 0);
       cout << "Done printing tree" << endl;
     }
   }
+
+  /*
+  if (nExperiences % 100 == 0){
+    cout << endl << "DT: " << id << endl;
+    printTree(root, 0);
+    cout << "Done printing tree" << endl;
+    
+    // look at error
+    float errorSum = 0.0;
+    for (int i = 0; i < nExperiences; i++){
+      e = experiences[i];
+      std::map<float, float> answer;
+      testInstance(e->input, &answer);
+      float val = answer.begin()->first;
+      float error = fabs(val - e->output);
+      errorSum += error;
+    }
+    float avgError = errorSum / (float)nExperiences;
+    cout << "avgError: " << avgError << endl << endl;
+  }
+  */
 
   return modelChanged;
 
@@ -269,14 +277,9 @@ bool M5Tree::trainInstance(classPair &instance){
 
 
 // here the target output will be a single value
-bool M5Tree::trainInstances(std::vector<classPair> &instances){
-  if (DTDEBUG) cout << id << " DT trainInstances: " 
-                            << instances.size() 
-                            << " nExp: " << nExperiences << endl;
-  nfeat = instances[0].in.size();
-  
-  //  featPct *= 0.99;
-  
+bool LinearSplitsTree::trainInstances(std::vector<classPair> &instances){
+  if (DTDEBUG) cout << "DT trainInstances: " << instances.size() << endl;
+
   bool modelChanged = false;
 
   bool doBuild = false;
@@ -289,7 +292,7 @@ bool M5Tree::trainInstances(std::vector<classPair> &instances){
 
     // take from static array until we run out
     tree_experience *e;
-    if (nExperiences < N_M5_EXP){
+    if (nExperiences < N_LST_EXP){
       // from statically created set of experiences
       e = &(allExp[nExperiences]);
 
@@ -325,6 +328,27 @@ bool M5Tree::trainInstances(std::vector<classPair> &instances){
       }
       cout << endl << " Now have " << nExperiences << " experiences." << endl;
     }
+
+    /*
+    if (nExperiences % 100 == 0){
+      cout << endl << "DT: " << id << endl;
+      printTree(root, 0);
+      cout << "Done printing tree" << endl;
+      
+      // look at error
+      float errorSum = 0.0;
+      for (int i = 0; i < nExperiences; i++){
+        e = experiences[i];
+        std::map<float, float> answer;
+        testInstance(e->input, &answer);
+        float val = answer.begin()->first;
+        float error = fabs(val - e->output);
+        errorSum += error;
+      }
+      float avgError = errorSum / (float)nExperiences;
+      cout << "avgError: " << avgError << endl << endl;
+    }
+    */
 
     // depending on mode/etc, maybe re-build tree
 
@@ -370,9 +394,9 @@ bool M5Tree::trainInstances(std::vector<classPair> &instances){
   }
 
   if (modelChanged){
-    if (DTDEBUG) cout << "DT " << id << " tree re-built." << endl;
+    if (DTDEBUG || SPLITDEBUG) cout << "DT " << id << " tree re-built." << endl;
 
-    if (DTDEBUG){
+    if (DTDEBUG || SPLITDEBUG){
       cout << endl << "DT: " << id << endl;
       printTree(root, 0);
       cout << "Done printing tree" << endl;
@@ -384,17 +408,21 @@ bool M5Tree::trainInstances(std::vector<classPair> &instances){
 }
 
 
-void M5Tree::rebuildTree(){
+void LinearSplitsTree::rebuildTree(){
   //cout << "rebuild tree " << id << " on exp: " << nExperiences << endl;
-  //  deleteTree(root);
+  //deleteTree(root);
+
+  // re-calculate avg error for root
+  root->avgError = calcAvgErrorforSet(experiences);
+
   buildTree(root, experiences, false);
   //cout << "tree " << id << " rebuilt. " << endl;
 }
 
 
 // TODO: here we want to return the probability of the output value being each of the possible values, in the stochastic case
-void M5Tree::testInstance(const std::vector<float> &input, std::map<float, float>* retval){
-  if (DTDEBUG) cout << "testInstance" << endl;
+void LinearSplitsTree::testInstance(const std::vector<float> &input, std::map<float, float>* retval){
+  if (DTDEBUG) cout << "testInstance on tree " << id << endl;
 
   retval->clear();
 
@@ -413,7 +441,7 @@ void M5Tree::testInstance(const std::vector<float> &input, std::map<float, float
 
 }
 
-float M5Tree::getConf(const std::vector<float> &input){
+float LinearSplitsTree::getConf(const std::vector<float> &input){
   if (DTDEBUG) cout << "numVisits" << endl;
 
   // in case the tree is empty
@@ -421,9 +449,8 @@ float M5Tree::getConf(const std::vector<float> &input){
     return 0;
   }
 
-  if (lastNode == NULL){
+  if (lastNode == NULL)
     return 0;
-  }
 
   // follow through tree to leaf
   //tree_node* leaf = traverseTree(root, input);
@@ -441,8 +468,9 @@ float M5Tree::getConf(const std::vector<float> &input){
 // to get more info on potential splits
 
 
+
 // init the tree
-void M5Tree::initTree(){
+void LinearSplitsTree::initTree(){
   if (DTDEBUG) cout << "initTree()" << endl;
   root = allocateNode();
 
@@ -458,7 +486,7 @@ void M5Tree::initTree(){
 
 
 // init a tree node
-void M5Tree::initTreeNode(tree_node* node){
+void LinearSplitsTree::initTreeNode(tree_node* node){
   if (DTDEBUG) cout << "initTreeNode()";
 
   node->id = nnodes++;
@@ -467,7 +495,7 @@ void M5Tree::initTreeNode(tree_node* node){
   totalnodes++;
   if (totalnodes > maxnodes){
     maxnodes = totalnodes;
-    if (DTDEBUG) cout << id << " M5 MAX nodes: " << maxnodes << endl;
+    if (DTDEBUG) cout << id << " LS MAX nodes: " << maxnodes << endl;
   }
 
   // split criterion
@@ -476,7 +504,9 @@ void M5Tree::initTreeNode(tree_node* node){
 
   // current data
   node->nInstances = 0;
-  node->constant = 0;
+
+  // coefficients
+  node->constant = 0.0;
 
   // coefficients will get resized later
   //  node->coefficients.resize(2,0);
@@ -486,10 +516,12 @@ void M5Tree::initTreeNode(tree_node* node){
   node->r = NULL;
 
   node->leaf = true;
+  node->avgError = 10000;
 
 }
 
-void M5Tree::deleteTree(tree_node* node){
+/** delete current tree */
+void LinearSplitsTree::deleteTree(tree_node* node){
   if (DTDEBUG) cout << "deleteTree, node=" << node->id << endl;
 
   if (node==NULL)
@@ -499,7 +531,7 @@ void M5Tree::deleteTree(tree_node* node){
 
   node->nInstances = 0;
   node->coefficients.clear();
-
+ 
   //recursively call deleteTree on children
   // then delete them
   if (!node->leaf){
@@ -518,15 +550,17 @@ void M5Tree::deleteTree(tree_node* node){
     }
   }
 
-  node->leaf  = true;
   node->dim = -1;
+  node->leaf = true;
+  node->avgError = 10000;
   node->val = -1;
   node->constant = 0;
 }
 
 
-M5Tree::tree_node* M5Tree::getCorrectChild(tree_node* node,
-                                           const std::vector<float> &input){
+/** Get the correct child of this node based on the input */
+LinearSplitsTree::tree_node* LinearSplitsTree::getCorrectChild(tree_node* node,
+                                                               const std::vector<float> &input){
 
   if (DTDEBUG) cout << "getCorrectChild, node=" << node->id << endl;
 
@@ -537,8 +571,9 @@ M5Tree::tree_node* M5Tree::getCorrectChild(tree_node* node,
 
 }
 
-M5Tree::tree_node* M5Tree::traverseTree(tree_node* node,
-                                        const std::vector<float> &input){
+/** Traverse the tree to the leaf for this input. */
+LinearSplitsTree::tree_node* LinearSplitsTree::traverseTree(tree_node* node,
+                                                            const std::vector<float> &input){
 
   if (DTDEBUG) cout << "traverseTree, node=" << node->id << endl;
 
@@ -550,10 +585,12 @@ M5Tree::tree_node* M5Tree::traverseTree(tree_node* node,
 }
 
 
-bool M5Tree::passTest(int dim, float val, const std::vector<float> &input){
+/** Decide if this passes the test */
+bool LinearSplitsTree::passTest(int dim, float val, const std::vector<float> &input){
   if (DTDEBUG) cout << "passTest, dim=" << dim << ",val=" << val 
                     << ",input["<<dim<<"]=" << input[dim] <<endl;
 
+  
   if (input[dim] > val)
     return false;
   else
@@ -562,7 +599,8 @@ bool M5Tree::passTest(int dim, float val, const std::vector<float> &input){
 }
 
 
-void M5Tree::buildTree(tree_node *node,
+/** Build the tree from this node down using this set of experiences. */
+void LinearSplitsTree::buildTree(tree_node *node,
                        const std::vector<tree_experience*> &instances,
                        bool changed){
   if(DTDEBUG) cout << "buildTree, node=" << node->id
@@ -581,6 +619,7 @@ void M5Tree::buildTree(tree_node *node,
   // first, add instances to tree
   node->nInstances = instances.size();
 
+  // add each output to this node
   bool allSame = true;
   float val0 = instances[0]->output;
   for (unsigned i = 1; i < instances.size(); i++){
@@ -592,79 +631,51 @@ void M5Tree::buildTree(tree_node *node,
 
   // see if they're all the same
   if (allSame){
-    makeLeaf(node);
-    node->constant = instances[0]->output;
+    makeLeaf(node, instances);
     if (DTDEBUG){
       cout << "Tree " << id << " node " << node->id 
            << " All " << node->nInstances
            << " classified with output "
-           << instances[0]->output << endl;
+           << instances[0]->output << ", " << node->constant << endl;
     }
     return;
   }
 
-  // check if this is a leaf and linear model has no error
-  if (node->leaf && node->coefficients.size() > 0) {
-    //cout << "Node " << node->id << " is leaf, checking lm" << endl;
-    float errorSum = 0;
-    for (unsigned i = 0; i < instances.size(); i++){
-      // get prediction for instance and compare with actual output
-      tree_node* leaf = traverseTree(node, instances[i]->input);
-      std::map<float, float> retval;
-      leafPrediction(leaf, instances[i]->input, &retval);
-      float prediction = retval.begin()->first;
-      float absError = fabs(prediction - instances[i]->output);
-      errorSum += absError;
-      //cout << "instance " << i << " leaf predicted: " << prediction
-      //     << " actual: " << instances[i]->output
-      //   << " error: " << absError << endl;
-    }
-    float avgError = errorSum / (float)instances.size();
-    if (avgError < 0.001){
-      //      cout << "stick with linear model" << endl;
+  // check if linear model has no error
+  if (node != root && node->avgError < MIN_ER){
+    makeLeaf(node, instances);
+    if (node->avgError < MIN_ER){
+      if (DTDEBUG) {
+      cout << "Tree " << id << " node " << node->id
+           << " has low error " << node->avgError << " keeping as leaf" << endl;
+      }
       return;
     }
   }
-
-  // if not, calculate SDR to determine best split
-  if (SPLITDEBUG) cout << endl << "Creating new decision node" << endl;
+  
+  // if not, calculate ER to determine best split
+  if (SPLITDEBUG) cout << endl << "Creating new decision node " << id << "-" << node->id << endl;
 
   node->leaf = false;
   //node->nInstances++;
 
-  float bestSDR = -1.0;
+  float bestER = -1.0;
   int bestDim = -1;
   float bestVal = -1;
   std::vector<tree_experience*> bestLeft;
   std::vector<tree_experience*> bestRight;
+  float leftError = 10000;
+  float rightError = 10000;
 
-  testPossibleSplits(instances, &bestSDR, &bestDim, &bestVal, &bestLeft, &bestRight);
+  testPossibleSplits(node->avgError, instances, &bestER, &bestDim, &bestVal, &bestLeft, &bestRight, &leftError, &rightError);
 
-  implementSplit(node, instances, bestSDR, bestDim, bestVal, bestLeft, bestRight, changed);
-
-  // possibly replace split node with linear regression model
-  pruneTree(node, instances);
-
-}
-
-
-void M5Tree::makeLeaf(tree_node* node){
-
-  removeChildren(node);
-
-  // make sure there are enough coefficients for all the features
-  // and that they are 0
-  if (node->coefficients.size() != (unsigned)nfeat){
-    node->coefficients.resize(nfeat, 0);
-  }
-  for (unsigned i = 0; i < node->coefficients.size(); i++){
-    node->coefficients[i] = 0;
-  }
+  implementSplit(node, instances, bestER, bestDim, bestVal, bestLeft, bestRight, changed, leftError, rightError);
 
 }
 
 
-void M5Tree::removeChildren(tree_node* node){
+void LinearSplitsTree::makeLeaf(tree_node* node, const std::vector<tree_experience*> &instances){
+
   // check on children
   if (node->l != NULL){
     deleteTree(node->l);
@@ -679,150 +690,53 @@ void M5Tree::removeChildren(tree_node* node){
   }
 
   node->leaf = true;
-}
 
-
-void M5Tree::pruneTree(tree_node *node,
-                       const std::vector<tree_experience*> &instances){
-  if (LMDEBUG || DTDEBUG){
-    printTree(root, 0);
-    cout << "pruneTree, node=" << node->id
-         << ",nInstances:" << instances.size() << endl;
-  }
-
-  // TODO: no pruning right now
-  //  return;
-
-  // calculate error of current subtree
-  float subtreeErrorSum = 0;
-  for (unsigned i = 0; i < instances.size(); i++){
-    
-    // get prediction for instance and compare with actual output
-    tree_node* leaf = traverseTree(node, instances[i]->input);
-    std::map<float, float> retval;
-    leafPrediction(leaf, instances[i]->input, &retval);
-    float prediction = retval.begin()->first;
-    float absError = fabs(prediction - instances[i]->output);
-    subtreeErrorSum += absError;
-    if (LMDEBUG || DTDEBUG){
-      cout << "instance " << i << " subtree predicted: " << prediction
-           << " actual: " << instances[i]->output
-           << " error: " << absError << endl;
-    }
-  }
-  if (instances.size() < 3){
-    if (LMDEBUG || DTDEBUG) cout << "instances size <= 2!!!" << endl;
-    return;
-  }
-
-  float avgTreeError = subtreeErrorSum / (float)instances.size();
-
-  // if this is zero error, we're not going to replace it
-  if (false && avgTreeError <= 0.0001){
-    if (LMDEBUG || DTDEBUG) 
-      cout << "Sub-tree is perfect (" << avgTreeError
-           << "), do not attempt LM" << endl;
-    return;
-  }
-                                                                
-  // figure out tree feats used
-  std::vector<bool> treeFeatsUsed(instances[0]->input.size(), false);
-  getFeatsUsed(node, &treeFeatsUsed);
-  int nTreeFeatsUsed = 0;
-  for (unsigned i = 0; i < treeFeatsUsed.size(); i++){
-    if (treeFeatsUsed[i])
-      nTreeFeatsUsed++;
-  }
-
-  // Just use them all... otherwise we ignore some that weren't good enough
-  // for splitting, but are good here
-  std::vector<bool> featsUsed(instances[0]->input.size(), true);
-  int nFeatsUsed = featsUsed.size();
-
-  // or just use ones allowed by subtree
-  if (!ALLOW_ALL_FEATS){
-    featsUsed = treeFeatsUsed;
-    nFeatsUsed = nTreeFeatsUsed;
-  }
-
-  // no features to build linear model on
-  if (nFeatsUsed == 0){
-    if (LMDEBUG || DTDEBUG) cout << "No features for LM" << endl;
-    return;
-  }
-
-  // add on error bonus based on nInstances and nParams
-  float treeErrorEst = avgTreeError;
-  float denom = (instances.size() - nTreeFeatsUsed);
-  if (denom < 1){
-    denom = 0.5;
-    if (LMDEBUG) {
-      cout << "denom of tree error factor is " << denom 
-           << " with nInst " << instances.size() 
-           << " nfeats: " << nTreeFeatsUsed << endl;
-    }
-  }
-  treeErrorEst *= (instances.size() + nTreeFeatsUsed) / denom;
-
-  // fit linear model to this set of instances
-  float lmErrorSum = 0;
-  int nlmFeats = 0;
+  // fit linear model
   if (SIMPLE)
-    nlmFeats = fitSimpleLinearModel(node, instances, featsUsed, nFeatsUsed, &lmErrorSum);
+    node->avgError = fitSimpleLinearModel(instances, &(node->constant), &(node->coefficients));
   else 
-    nlmFeats = fitLinearModel(node, instances, featsUsed, nFeatsUsed, &lmErrorSum);
+    node->avgError = fitMultiLinearModel(instances, &(node->constant), &(node->coefficients));
 
-  float avgLMError = lmErrorSum / (float)instances.size();
-
-  float lmErrorEst = avgLMError;
-  float denom2 = (instances.size() - nlmFeats);
-  if (denom2 < 1){
-    denom2 = 0.5;
-    if (LMDEBUG) {
-      cout << "denom2 of lm error factor is " << denom2
-           << " with nInst " << instances.size() 
-           << " nfeats: " << nlmFeats << endl;
-    }
-  }
-  lmErrorEst *= (instances.size() + nlmFeats) / denom2; 
-
-  // replace subtree with linear model?
-  if (LMDEBUG || DTDEBUG) {
-    cout << "Sub-Tree Error: " << treeErrorEst << ", lm Error: "
-         << lmErrorEst << endl;
-  }
-  //  if (lmErrorEst < (treeErrorEst + 0.0001)){
-  if (lmErrorEst < (treeErrorEst + 0.1*MIN_SDR)){
-  //if (lmErrorEst < treeErrorEst){
-    if (LMDEBUG || DTDEBUG)
-      cout << node->id << " replace tree with linear model" << endl;
-    removeChildren(node);
-  } else {
-    // remove coefficients again, for memory
-    //    node->coefficients.clear();
+  if (DTDEBUG || LMDEBUG){
+    cout << "make leaf, fit linear model with constant " << node->constant 
+         << "  error: " << node->avgError << endl;
   }
 
 }
 
 
-int M5Tree::fitLinearModel(tree_node *node,
-                           const std::vector<tree_experience*> &instances,
-                           std::vector<bool> featureMask,
-                           int nFeats, float* resSum){
-  
-  if(DTDEBUG || LMDEBUG) cout << "fitLinearModel, node=" << node->id
+float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> &instances,
+                                            float *bestConstant, 
+                                            std::vector<float> *bestCoefficients){
+  if(DTDEBUG || LMDEBUG) cout << "fitMultiLinearModel"
                               << ",nInstances:" << instances.size() << endl;
 
   // make sure there are enough coefficients for all the features
-  if (node->coefficients.size() != instances[0]->input.size()){
-    node->coefficients.resize(instances[0]->input.size(), 0);
+  if (bestCoefficients->size() != instances[0]->input.size()){
+    bestCoefficients->resize(instances[0]->input.size(), 0);  
   }
 
-  node->constant = 0.0;
+  // in case of size 1
+  if (instances.size() == 1){
+    *bestConstant = instances[0]->output;
+    for (unsigned i = 0; i < bestCoefficients->size(); i++){
+      (*bestCoefficients)[i] = 0.0;
+    }
+    if (LMDEBUG || SPLITDEBUG){
+      cout << "  Singleton constant: " 
+           << *bestConstant << " avgE: " << 0 << endl;
+    }
+    return 0;
+  }
+
+  // loop through all features, try simple single variable regression
+  // keep track of error/coefficient of each
+  float bestError = 100000;
+  float avgError = 100000;
   bool doRegression = true;
   int ntimes = 0;
-  int nlmFeats = 10;
-  (*resSum) = 1000000;
+
+  std::vector<bool> featureMask(bestCoefficients->size(), true);
 
   while (doRegression){
     //cout << id << " Attempt linear model " << ntimes << endl;
@@ -830,7 +744,12 @@ int M5Tree::fitLinearModel(tree_node *node,
 
     int nObs = (int)instances.size();
 
-    //cout << "with nObs: " << nObs << " and nFeats: " << nFeats << endl;
+    int nFeats = 0;
+    for (unsigned i = 0; i < featureMask.size(); i++){
+      if (featureMask[i]) nFeats++;
+    }
+    if (nFeats < 1)
+      break;
 
     // no feats or obs, no model to build
     if (nObs == 0 || nFeats == 0)
@@ -852,24 +771,13 @@ int M5Tree::fitLinearModel(tree_node *node,
       // go through all features
       int featIndex = 1;
       for (unsigned j = 0; j < featureMask.size(); j++){
-        node->coefficients[j] = 0;
+        (*bestCoefficients)[j] = 0;
         if (!featureMask[j])
           continue;
-      
+
         if (constants[j] && e->input[j] != instances[0]->input[j]){
           constants[j] = false;
         }
-
-	/*
-        if (rng.uniform() < featPct){
-          featureMask[j] = false;
-          nFeats--;
-          if (nFeats > 0)
-            continue;
-          else
-            break;
-        }
-	*/
 
         if (i == nObs-1 && constants[j]){
           //cout << "PROBLEM: feat " << j << " is constant!" << endl;
@@ -896,7 +804,7 @@ int M5Tree::fitLinearModel(tree_node *node,
       if (LMDEBUG) cout << " out: " << e->output << endl;
     }
 
-    if (foundProblem || nFeats == 0)
+    if (foundProblem)
       continue;
 
     // make vector of 1s
@@ -914,13 +822,13 @@ int M5Tree::fitLinearModel(tree_node *node,
     Real mval = Sum(Y) / nObs;
     YC = Y - Ones * mval;
 
+    // form sum of squares and product matrix
+    //    [use << rather than = for copying Matrix into SymmetricMatrix]
+    SymmetricMatrix SSQ;
+    SSQ << XC.t() * XC;
+
     Try {
 
-      // form sum of squares and product matrix
-      //    [use << rather than = for copying Matrix into SymmetricMatrix]
-      SymmetricMatrix SSQ;
-      SSQ << XC.t() * XC;
-      
       ///////////////////////////
       // Cholesky Method
       LowerTriangularMatrix L = Cholesky(SSQ);
@@ -931,12 +839,13 @@ int M5Tree::fitLinearModel(tree_node *node,
       
       //////////////////////////
       // Least Squares Method
+      /*
       // calculate estimate
       //    [bracket last two terms to force this multiplication first]
       //    [ .i() means inverse, but inverse is not explicity calculated]
-      //ColumnVector A = SSQ.i() * (XC.t() * YC);
+      ColumnVector A = SSQ.i() * (XC.t() * YC);
       //////////////////////////
-  
+      */
   
       // calculate estimate of constant term
       //    [AsScalar converts 1x1 matrix to Real]
@@ -950,28 +859,24 @@ int M5Tree::fitLinearModel(tree_node *node,
 
       // print out answers
       // for each instance
-      (*resSum) = 0;
+      bestError = 0;
       for (int i = 0; i < nObs; i++){
         if (DTDEBUG || LMDEBUG){
           cout << "instance " << i << " linear model predicted: " << Fitted(i+1)
                << " actual: " << instances[i]->output
                << " error: " << Residual(i+1) << endl;
         }
-        (*resSum) += fabs(Residual(i+1));
+        bestError += fabs(Residual(i+1));
       }
-
+      avgError = bestError / (float)nObs;
 
       // coeff
-      nlmFeats = 0;
       for (int i = 0; i < nFeats; i++){
         if (DTDEBUG || LMDEBUG) cout << "Coeff " << i << " on feat: " << featIndices[i] << " is " << A(i+1) << endl;
-        node->coefficients[featIndices[i]] = A(i+1);
-        if (A(i+1) != 0)
-          nlmFeats++;
+        (*bestCoefficients)[featIndices[i]] = A(i+1);
       }
       if (DTDEBUG || LMDEBUG) cout << "constant is " << a << endl;
-      node->constant = a;
-
+      *bestConstant = a;
 
     }
 
@@ -981,8 +886,31 @@ int M5Tree::fitLinearModel(tree_node *node,
       //cout << ntimes << " linear regression had exception" << endl;
       //<< BaseException::what() <<endl;
 
+      /*
+      for (int i = 0; i < nObs; i++){
+        tree_experience *e = instances[i];
+        cout << "Obs: " << i;
+        
+        // go through all features
+        int featIndex = 1;
+        for (unsigned j = 0; j < featureMask.size(); j++){
+          node->coefficients[j] = 0;
+          if (!featureMask[j])
+            continue;
+          
+          
+          cout << " Feat " << featIndex << " index " << j
+               << " val " << X(i+1,featIndex) << ",";
+          
+          featIndex++;
+        }
+        
+        cout << " out: " << e->output << endl;
+      }
+      */
+
       // tried this already, stop now
-      if (ntimes > 1 || nFeats < 2){
+      if (ntimes > 2 || nFeats < 2){
         //cout << "max regression" << endl;
         doRegression = false;
         break;
@@ -996,36 +924,50 @@ int M5Tree::fitLinearModel(tree_node *node,
         }
       }
       continue;
-    }
+    } // catch
 
     // it worked, dont need to do it again
     doRegression = false;
 
   }
 
-  // return # features used
-  return nlmFeats;
+  // return error
+  return avgError;
 
 }
 
-int M5Tree::fitSimpleLinearModel(tree_node *node,
-                                 const std::vector<tree_experience*> &instances,
-                                 std::vector<bool> featureMask,
-                                 int nFeats, float* resSum){
-  if(DTDEBUG || LMDEBUG) cout << "fitSimpleLinearModel, node=" << node->id
+
+float LinearSplitsTree::fitSimpleLinearModel(const std::vector<tree_experience*> &instances,
+                                             float *bestConstant, std::vector<float> *bestCoefficients){
+  if(DTDEBUG || LMDEBUG || SPLITDEBUG) cout << " fitSimpleLinearModel, "
                               << ",nInstances:" << instances.size() << endl;
   
   // make sure there are enough coefficients for all the features
-  if (node->coefficients.size() != instances[0]->input.size()){
-    node->coefficients.resize(instances[0]->input.size(), 0);
+  if (bestCoefficients->size() != instances[0]->input.size()){
+    bestCoefficients->resize(instances[0]->input.size(), 0);  
   }
 
   // loop through all features, try simple single variable regression
   // keep track of error/coefficient of each
   int bestFeat = -1;
   float bestCoeff = -1;
-  float bestError = 1000000;
-  float bestConstant = -1;
+  float bestError = 100000;
+  *bestConstant = -1;
+
+  // in case of size 1
+  if (instances.size() == 1){
+    bestFeat = 0;
+    *bestConstant = instances[0]->output;
+    for (unsigned i = 0; i < bestCoefficients->size(); i++){
+      (*bestCoefficients)[i] = 0.0;
+    }
+    bestError = 0;
+    if (LMDEBUG || SPLITDEBUG){
+      cout << "  Singleton constant: " 
+           << *bestConstant << " avgE: " << bestError << endl;
+    }
+    return 0;
+  }
 
   std::vector<float> xsum(instances[0]->input.size(), 0);
   std::vector<float> xysum(instances[0]->input.size(), 0);
@@ -1039,7 +981,6 @@ int M5Tree::fitSimpleLinearModel(tree_node *node,
     
     // go through all features
     for (unsigned j = 0; j < instances[0]->input.size(); j++){
-      if (!featureMask[j]) continue;
       if (LMDEBUG) cout << ", F" << j << ": " << e->input[j];
       xsum[j] += e->input[j];
       xysum[j] += (e->input[j] * e->output);
@@ -1048,21 +989,23 @@ int M5Tree::fitSimpleLinearModel(tree_node *node,
     ysum += e->output;
     if (LMDEBUG) cout << ", out: " << e->output << endl;
   }
-  
+
   // now go through all features and calc coeff and constant
   for (unsigned j = 0; j < instances[0]->input.size(); j++){
-    if (!featureMask[j]) continue;
-    /*
-      if (rng.uniform() < featPct){
-      continue;
-      }
-    */
     float coeff = (xysum[j] - xsum[j]*ysum/nObs)/(x2sum[j]-(xsum[j]*xsum[j])/nObs);
     float constant = (ysum/nObs) - coeff*(xsum[j]/nObs);
-    
+
+    // so we don't get absurd coefficients that get rounded off earlier
+    if (fabs(coeff) < 1e-5){
+      coeff = 0.0;
+    }
+    if (fabs(constant) < 1e-5){
+      constant = 0.0;
+    }
+
     if (LMDEBUG) {
       cout << "Feat " << j << " coeff: " << coeff << ", constant: " 
-           << constant << " mask: " << featureMask[j] << endl;
+           << constant  << endl;
     }
 
     // now try to make predictions and see what error is
@@ -1074,92 +1017,85 @@ int M5Tree::fitSimpleLinearModel(tree_node *node,
       if (LMDEBUG) cout << "Instance " << i << " error: " << error << endl;
       errorSum += error;
     }
-    if (LMDEBUG) cout << "eSum: " << errorSum << endl;
+    float avgError = errorSum / (float)nObs;
+    if (LMDEBUG) cout << "avgError: " << avgError << endl;
 
     // check if this is the best
-    if (errorSum < bestError){
-      bestError = errorSum;
+    if (avgError < bestError){
+      bestError = avgError;
       bestFeat = j;
-      bestConstant = constant;
+      *bestConstant = constant;
       bestCoeff = coeff;
     }
   }
   
-  if (LMDEBUG){
-    cout << "SLM feat: " << bestFeat << " coeff: " 
+  if (LMDEBUG || SPLITDEBUG){
+    cout << "  LST feat: " << bestFeat << " coeff: " 
          << bestCoeff << " constant: " 
-         << bestConstant << " avgE: " << (bestError/nObs) << endl;
+         << *bestConstant << " avgE: " << bestError << endl;
   }
 
   if (bestFeat < 0 || bestFeat > (int)instances[0]->input.size()){
-    node->constant = 0.0;
-    for (unsigned i = 0; i < node->coefficients.size(); i++){
-      node->coefficients[i] = 0.0;
+    for (unsigned i = 0; i < bestCoefficients->size(); i++){
+      (*bestCoefficients)[i] = 0.0;
     }
-    (*resSum) = 1000000;
-    return 10;
+    *bestConstant = 0.0;
+    return 100000;
   }
 
-  // pick best feature somehow
-  node->constant = bestConstant;
-  for (unsigned i = 0; i < node->coefficients.size(); i++){
-    if (i == (unsigned)bestFeat){
-      node->coefficients[i] = bestCoeff;
-      if (LMDEBUG) cout << "Set coefficient on feat " << i << " to " << bestCoeff << endl;
-    }
-    else {
-      node->coefficients[i] = 0.0;
-    }
+  // fill in coeff vector
+  for (unsigned i = 0; i < bestCoefficients->size(); i++){
+    if (i == (unsigned)bestFeat)
+      (*bestCoefficients)[i] = bestCoeff;
+    else
+      (*bestCoefficients)[i] = 0.0; 
   }
 
-  (*resSum) = bestError;
-
-  // only use 1 input feature this way
-  return 1;
+  return bestError;
  
 }
 
 
 
 
-void M5Tree::implementSplit(tree_node* node, 
-                            const std::vector<tree_experience*> &instances,
-                            float bestSDR, int bestDim,
-                            float bestVal, 
-                            const std::vector<tree_experience*> &bestLeft,
-                            const std::vector<tree_experience*> &bestRight,
-                            bool changed){
+void LinearSplitsTree::implementSplit(tree_node* node, 
+                                      const std::vector<tree_experience*> &instances,
+                                      float bestER, int bestDim,
+                                      float bestVal, 
+                                      const std::vector<tree_experience*> &bestLeft,
+                                      const std::vector<tree_experience*> &bestRight,
+                                      bool changed, float leftError, float rightError){
   if (DTDEBUG) cout << "implementSplit node=" << node->id
-                    << ",sdr=" << bestSDR
+                    << ",er=" << bestER
                     << ",dim=" << bestDim
                     << ",val=" << bestVal 
                     << ",chg=" << changed << endl;
 
 
   // see if this should still be a leaf node
-  if (bestSDR < MIN_SDR){
-    makeLeaf(node);
-    float valSum = 0;
-    for (unsigned i = 0; i < instances.size(); i++){
-      valSum += instances[i]->output;
-    }
-    float avg = valSum / (float)(instances.size());
-    node->constant = avg;
+  if (bestER < MIN_ER){
+    makeLeaf(node, instances);
+
     if (SPLITDEBUG || STOCH_DEBUG){
-      cout << "DT " << id << " Node " << node->id << " Poor sdr "
+      cout << " DT " << id << " Node " << node->id << " Poor er "
            << node->nInstances
            << " instances classified at leaf " << node->id
-           << " with multiple outputs " << endl;
+           << " with er " << bestER << " constant: " << node->constant << endl;
     }
     return;
   }
 
+  //cout << id << " implement split with er " << bestER << endl;
+
   // see if this split changed or not
   // assuming no changes above
   if (!changed && node->dim == bestDim && node->val == bestVal
-       && !node->leaf && node->l != NULL && node->r != NULL){
+      && !node->leaf && node->l != NULL && node->r != NULL){
     // same split as before.
-    if (DTDEBUG || SPLITDEBUG) cout << "Same split as before" << endl;
+    if (DTDEBUG || SPLITDEBUG) cout << "Same split as before " << node->id << endl;
+    node->l->avgError = leftError;
+    node->r->avgError = rightError;
+
 
     // see which leaf changed
     if (bestLeft.size() > (unsigned)node->l->nInstances){
@@ -1185,7 +1121,7 @@ void M5Tree::implementSplit(tree_node* node,
 
   if (SPLITDEBUG) cout << "Best split was cut with val " << node->val
                        << " on dim " << node->dim
-                       << " with sdr: " << bestSDR << endl;
+                       << " with er: " << bestER << endl;
 
   if (DTDEBUG) cout << "Left has " << bestLeft.size()
                     << ", right has " << bestRight.size() << endl;
@@ -1196,7 +1132,7 @@ void M5Tree::implementSplit(tree_node* node,
          << " right: " << bestRight.size() << endl;
     cout << "Split was cut with val " << node->val
          << " on dim " << node->dim
-         << " with sdr: " << bestSDR << endl;
+         << " with er: " << bestER << endl;
     exit(-1);
   }
 
@@ -1213,50 +1149,22 @@ void M5Tree::implementSplit(tree_node* node,
 
   // recursively build the sub-trees to this one
   if (DTDEBUG) cout << "Building left tree for node " << node->id << endl;
+  node->l->avgError = leftError;
   buildTree(node->l, bestLeft, true);
   if (DTDEBUG) cout << "Building right tree for node " << node->id << endl;
+  node->r->avgError = rightError;
   buildTree(node->r, bestRight, true);
 
 }
 
 
-void M5Tree::getFeatsUsed(tree_node* node, std::vector<bool> *featsUsed){
-
-  // if leaf, ones from linear model
-  if (node->leaf){
-    //cout << "coeff size: " << node->coefficients.size() << endl;
-    for (unsigned i = 0; i < node->coefficients.size(); i++){
-      if (node->coefficients[i] != 0){
-        if (LMDEBUG || DTDEBUG) cout << "Leaf node, used coeff " << i << endl;
-        (*featsUsed)[i] = true;
-      }
-    }
-    return;
-  }
-
-  // otherwise see what split was used
-  // and call for left and right sub-trees
-  (*featsUsed)[node->dim] = true;
-  if (LMDEBUG || DTDEBUG) cout << "Split node, used feat " << node->dim << endl;
-
-  getFeatsUsed(node->l, featsUsed);
-  getFeatsUsed(node->r, featsUsed);
-
-  return;
-}
-
-
-void M5Tree::testPossibleSplits(const std::vector<tree_experience*> &instances,
-                                float *bestSDR, int *bestDim,
-                                float *bestVal, 
-                                std::vector<tree_experience*> *bestLeft,
-                                std::vector<tree_experience*> *bestRight) {
-  if (DTDEBUG) cout << "testPossibleSplits" << endl;
-
-
-  // calculate sd for the set
-  float sd = calcSDforSet(instances);
-  //if (DTDEBUG) cout << "I: " << I << endl;
+void LinearSplitsTree::testPossibleSplits(float avgError, const std::vector<tree_experience*> &instances,
+                                          float *bestER, int *bestDim,
+                                          float *bestVal, 
+                                          std::vector<tree_experience*> *bestLeft,
+                                          std::vector<tree_experience*> *bestRight,
+                                          float *bestLeftError, float *bestRightError) {
+  if (DTDEBUG || SPLITDEBUG) cout << "testPossibleSplits, error=" << avgError << endl;
 
   int nties = 0;
 
@@ -1277,21 +1185,25 @@ void M5Tree::testPossibleSplits(const std::vector<tree_experience*> &instances,
       // here (decision is taken from the random set that are left)
       if (rng.uniform() < featPct)
         continue;
-
+      
       std::vector<tree_experience*> left;
       std::vector<tree_experience*> right;
+      float leftError = 10000;
+      float rightError = 10000;
 
       // splits that are cuts
       float splitval = (*j);
-      float sdr = calcSDR(idim, splitval, instances, sd, left, right);
+      float er = calcER(idim, splitval, instances, avgError, left, right, &leftError, &rightError);
 
-      if (SPLITDEBUG) cout << " CUT split val " << splitval
-                           << " on dim: " << idim << " had sdr "
-                           << sdr << endl;
+      if (SPLITDEBUG){
+        cout << id << " CUT split val " << splitval
+             << " on dim: " << idim << " had er "
+             << er << endl;
+      }
 
-      // see if this is the new best sdr
-      compareSplits(sdr, idim, splitval, left, right, &nties,
-                    bestSDR, bestDim, bestVal, bestLeft, bestRight);
+      // see if this is the new best er
+      compareSplits(er, idim, splitval, left, right, &nties, leftError, rightError,
+                    bestER, bestDim, bestVal,bestLeft, bestRight, bestLeftError, bestRightError);
 
 
     } // j loop
@@ -1300,21 +1212,24 @@ void M5Tree::testPossibleSplits(const std::vector<tree_experience*> &instances,
 
 
 
-void M5Tree::compareSplits(float sdr, int dim, float val, 
-                           const std::vector<tree_experience*> &left, 
-                           const std::vector<tree_experience*> &right,
-                           int *nties, float *bestSDR, int *bestDim,
-                           float *bestVal, 
-                           std::vector<tree_experience*> *bestLeft,
-                           std::vector<tree_experience*> *bestRight){
-  if (DTDEBUG) cout << "compareSplits sdr=" << sdr << ",dim=" << dim
+/** Decide if this split is better. */
+void LinearSplitsTree::compareSplits(float er, int dim, float val, 
+                                     const std::vector<tree_experience*> &left, 
+                                     const std::vector<tree_experience*> &right,
+                                     int *nties, float leftError, float rightError,
+                                     float *bestER, int *bestDim,
+                                     float *bestVal, 
+                                     std::vector<tree_experience*> *bestLeft,
+                                     std::vector<tree_experience*> *bestRight,
+                                     float *bestLeftError, float *bestRightError){
+  if (DTDEBUG) cout << "compareSplits er=" << er << ",dim=" << dim
                     << ",val=" << val  <<endl;
 
 
   bool newBest = false;
 
   // if its a virtual tie, break it randomly
-  if (fabs(*bestSDR - sdr) < SPLIT_MARGIN){
+  if (fabs(*bestER - er) < SPLIT_MARGIN){
     //cout << "Split tie, making random decision" << endl;
 
     (*nties)++;
@@ -1332,7 +1247,7 @@ void M5Tree::compareSplits(float sdr, int dim, float val,
   }
 
   // if its clearly better, set this as the best split
-  else if (sdr > *bestSDR){
+  else if (er > *bestER){
     newBest = true;
     *nties = 1;
   }
@@ -1340,27 +1255,31 @@ void M5Tree::compareSplits(float sdr, int dim, float val,
 
   // set the split features
   if (newBest){
-    *bestSDR = sdr;
+    *bestER = er;
     *bestDim = dim;
     *bestVal = val;
     *bestLeft = left;
     *bestRight = right;
+    *bestLeftError = leftError;
+    *bestRightError = rightError;
     if (SPLITDEBUG){
-      cout << "  New best sdr: " << *bestSDR
+      cout << "  New best er: " << *bestER
            << " with val " << *bestVal
            << " on dim " << *bestDim << endl;
     }
   } // newbest
 }
 
-float M5Tree::calcSDR(int dim, float val, 
-                      const std::vector<tree_experience*> &instances,
-                      float sd,
-                      std::vector<tree_experience*> &left,
-                      std::vector<tree_experience*> &right){
-  if (DTDEBUG) cout << "calcSDR, dim=" << dim
+/** Calculate error reduction for this possible split. */
+float LinearSplitsTree::calcER(int dim, float val,
+                               const std::vector<tree_experience*> &instances,
+                               float avgError,
+                               std::vector<tree_experience*> &left,
+                               std::vector<tree_experience*> &right,
+                               float *leftError, float *rightError){
+  if (DTDEBUG || SPLITDEBUG) cout << "calcER, dim=" << dim
                     << " val=" << val
-                    << " sd=" << sd
+                    << " err=" << avgError
                     << " nInstances= " << instances.size() << endl;
 
   left.clear();
@@ -1368,8 +1287,7 @@ float M5Tree::calcSDR(int dim, float val,
 
   // split into two sides
   for (unsigned i = 0; i < instances.size(); i++){
-    if (DTDEBUG) cout << "calcSDR - Classify instance " << i 
-                      << " on new split " << endl;
+    if (DTDEBUG) cout << " calcER - Classify instance " << i << " on new split " << endl;
 
     if (passTest(dim, val, instances[i]->input)){
       left.push_back(instances[i]);
@@ -1383,53 +1301,53 @@ float M5Tree::calcSDR(int dim, float val,
                     << ", right has " << right.size() << endl;
 
   // get sd for both sides
-  float sdLeft = calcSDforSet(left);
-  float sdRight = calcSDforSet(right);
+  *leftError = calcAvgErrorforSet(left);
+  *rightError = calcAvgErrorforSet(right);
 
   float leftRatio = (float)left.size() / (float)instances.size();
   float rightRatio = (float)right.size() / (float)instances.size();
+  float newError = (leftRatio * (*leftError) + rightRatio * (*rightError));
 
-  float sdr = sd - (leftRatio * sdLeft + rightRatio * sdRight);
+  float er = avgError - newError;
 
-  if (DTDEBUG){
-    cout << "LeftSD: " << sdLeft
-         << " RightSD: " << sdRight
-         << " SD: " << sd
-         << " SDR: " << sdr
+  if (DTDEBUG || SPLITDEBUG){
+    cout << "LeftError: " << *leftError
+         << " RightError: " << *rightError
+         << " NewError: " << newError
+         << " NodeError: " << avgError
+         << " ER: " << er
          << endl;
   }
 
-  return sdr;
+  return er;
 
 }
 
-float M5Tree::calcSDforSet(const std::vector<tree_experience*> &instances){
-  if (DTDEBUG) cout << "calcSDforSet" << endl;
+/** Calculate std deviation for set. */
+float LinearSplitsTree::calcAvgErrorforSet(const std::vector<tree_experience*> &instances){
+  if (DTDEBUG) cout << "calcAvgErrorforSet" << endl;
 
   int n = instances.size();
 
   if (n == 0)
     return 0;
 
-  double sum = 0;
-  double sumSqr = 0;
+  // fit a linear model to instances
+  // and figure out avg error of it
+  float constant;
+  float avgError = 0.0;
+  std::vector<float> coeff;
+  if (SIMPLE) 
+    avgError = fitSimpleLinearModel(instances, &constant, &coeff);
+  else 
+    avgError = fitMultiLinearModel(instances, &constant, &coeff);
 
-  // go through instances and calculate sums, sum of squares
-  for (unsigned i = 0; i < instances.size(); i++){
-    float val = instances[i]->output;
-    sum += val;
-    sumSqr += (val * val);
-  }
-
-  double mean = sum / (double)n;
-  double variance = (sumSqr - sum*mean)/(double)n;
-  float sd = sqrt(variance);
-
-  return sd;
-
+  return avgError;
 }
 
-std::set<float> M5Tree::getUniques(int dim, const std::vector<tree_experience*> &instances, float& minVal, float& maxVal){
+
+/** Returns the unique elements at this index */
+std::set<float> LinearSplitsTree::getUniques(int dim, const std::vector<tree_experience*> &instances, float& minVal, float& maxVal){
   if (DTDEBUG) cout << "getUniques,dim = " << dim;
 
   std::set<float> uniques;
@@ -1443,6 +1361,7 @@ std::set<float> M5Tree::getUniques(int dim, const std::vector<tree_experience*> 
     uniques.insert(instances[i]->input[dim]);
   }
 
+  
   // lets not try more than 100 possible splits per dimension
   if (uniques.size() > 100){
     float rangeInc = (maxVal - minVal) / 100.0;
@@ -1452,13 +1371,16 @@ std::set<float> M5Tree::getUniques(int dim, const std::vector<tree_experience*> 
     }
   }
   uniques.insert(maxVal);
+  
 
   if (DTDEBUG) cout << " #: " << uniques.size() << endl;
   return uniques;
 }
 
 
-float* M5Tree::sortOnDim(int dim, const std::vector<tree_experience*> &instances){
+/** Returns a list of the attributes in this dimension sorted
+    from lowest to highest. */
+float* LinearSplitsTree::sortOnDim(int dim, const std::vector<tree_experience*> &instances){
   if (DTDEBUG) cout << "sortOnDim,dim = " << dim << endl;
 
   float* values = new float[instances.size()];
@@ -1513,7 +1435,8 @@ float* M5Tree::sortOnDim(int dim, const std::vector<tree_experience*> &instances
 }
 
 
-void M5Tree::printTree(tree_node *t, int level){
+/** Print the tree for debug purposes. */
+void LinearSplitsTree::printTree(tree_node *t, int level){
 
   for (int i = 0; i < level; i++){
     cout << ".";
@@ -1550,7 +1473,8 @@ void M5Tree::printTree(tree_node *t, int level){
 
 
 // output a map of outcomes and their probabilities for this leaf node
-void M5Tree::leafPrediction(tree_node* leaf, const std::vector<float> &input, std::map<float, float>* retval){
+void LinearSplitsTree::leafPrediction(tree_node* leaf, const std::vector<float> &input,
+                                      std::map<float, float>* retval){
   if (DTDEBUG)
     cout << "Calculating output for leaf " << leaf->id << endl;
 
@@ -1569,12 +1493,13 @@ void M5Tree::leafPrediction(tree_node* leaf, const std::vector<float> &input, st
   if (DTDEBUG) cout << " prediction: " << prediction << endl;
 
   (*retval)[prediction] = 1.0;
+ 
 }
 
 
-void M5Tree::initNodes(){
+void LinearSplitsTree::initNodes(){
 
-  for (int i = 0; i < N_M5_NODES; i++){
+  for (int i = 0; i < N_LS_NODES; i++){
     initTreeNode(&(allNodes[i]));
     freeNodes.push_back(i);
     if (NODEDEBUG) 
@@ -1584,12 +1509,12 @@ void M5Tree::initNodes(){
 
 }
 
-M5Tree::tree_node* M5Tree::allocateNode(){
+LinearSplitsTree::tree_node* LinearSplitsTree::allocateNode(){
   if (freeNodes.empty()){
     tree_node* newNode = new tree_node;
     initTreeNode(newNode);
     if (NODEDEBUG) 
-      cout << "PROBLEM: No more pre-allocated nodes!!!" << endl
+      cout << id << " PROBLEM: No more pre-allocated nodes!!!" << endl
            << "return new node " << newNode->id 
            << ", now " << freeNodes.size() << " free nodes." << endl;
     return newNode;
@@ -1598,15 +1523,15 @@ M5Tree::tree_node* M5Tree::allocateNode(){
   int i = freeNodes.back();
   freeNodes.pop_back();
   if (NODEDEBUG) 
-    cout << "allocate node " << i << " with id " << allNodes[i].id 
+    cout << id << " allocate node " << i << " with id " << allNodes[i].id 
          << ", now " << freeNodes.size() << " free nodes." << endl;
   return &(allNodes[i]);
 }
 
-void M5Tree::deallocateNode(tree_node* node){
-  if (node->id >= N_M5_NODES){
+void LinearSplitsTree::deallocateNode(tree_node* node){
+  if (node->id >= N_LS_NODES){
     if (NODEDEBUG) 
-      cout << "dealloc extra node id " << node->id 
+      cout << id << " dealloc extra node id " << node->id 
            << ", now " << freeNodes.size() << " free nodes." << endl;
     delete node;
     return;
@@ -1614,6 +1539,6 @@ void M5Tree::deallocateNode(tree_node* node){
 
   freeNodes.push_back(node->id);
   if (NODEDEBUG) 
-    cout << "dealloc node " << node->id 
+    cout << id << " dealloc node " << node->id 
          << ", now " << freeNodes.size() << " free nodes." << endl;
 }
