+++ src/models/M5Tree.cc
@@ -1,116 +1,123 @@
-#include "LinearSplitsTree.hh"
+/** \file M5Tree.cc
+    Implements the M5 Decision tree, as described in:
+    "Learning with Continuous Classes" by J.R. Quinlan
+    "Inducing Model Trees for Continuous Classes" by Y. Wang and I.H. Witten
+    \author Todd Hester
+*/
 
-// LinearSplitsTree, from the following sources:
+#include "models/M5Tree.hh"
+
 
 // Include stuff for newmat matrix libraries
 
-#define WANT_MATH                    // include.h will get math fns
-                                     // newmatap.h will get include.h
-#include <newmat/newmatap.h>         // need matrix applications
+#define WANT_MATH                // include.h will get math fns
+                                 // newmatap.h will get include.h
+#include <newmat/newmatap.h>     // need matrix applications
 #ifdef use_namespace
-using namespace NEWMAT;              // access NEWMAT namespace
+using namespace NEWMAT;             // access NEWMAT namespace
 #endif
 
-// TODO:
-//  - save regression from split testing rather than re-building it later
 
 
-LinearSplitsTree::LinearSplitsTree(int id, int trainMode, int trainFreq, int m,
-                                   float featPct, bool simple, float min_er,
-                                   Random rng):
-  id(id), mode(trainMode), 
-  freq(trainFreq), M(m),
-  featPct(featPct), SIMPLE(simple), 
-  MIN_ER(min_er), rng(rng)
+M5Tree::M5Tree(int id, int trainMode, int trainFreq, int m,
+               float featPct, bool simple, bool allowAllFeats, 
+	       float min_sdr, Random rng):
+  id(id), mode(trainMode), freq(trainFreq), M(m),
+  featPct(featPct), SIMPLE(simple), ALLOW_ALL_FEATS(allowAllFeats),
+  MIN_SDR(min_sdr), rng(rng)
 {
 
   nnodes = 0;
   nOutput = 0;
   nExperiences = 0;
   hadError = false;
-  maxnodes = N_LS_NODES;
   totalnodes = 0;
+  maxnodes = N_M5_NODES;
+
 
   // how close a split has to be to be randomly selected
   SPLIT_MARGIN = 0.0; //0.02; //5; //01; //0.05; //0.2; //0.05;
 
-  LMDEBUG = false;// true;
-  DTDEBUG = false;//true;
-  SPLITDEBUG = false; //true;
+  LMDEBUG = false;
+  DTDEBUG = false;///true;
+  SPLITDEBUG = false;//true;
   STOCH_DEBUG = false; //true; //false; //true;
   INCDEBUG = false; //true; //false; //true;
   NODEDEBUG = false;
   COPYDEBUG = false; //true;
+  nfeat = 4;
 
-  cout << "Created linear splits decision tree " << id;
+  cout << "Created m5 decision tree " << id;
   if (SIMPLE) cout << " simple regression";
-  else cout << " multivariate regrssion";
+  else cout << " multivariate regression";
+  if (ALLOW_ALL_FEATS) cout << " (all feats)";
+  else cout << " (subtree feats)";
   if (DTDEBUG){
     cout << " mode: " << mode << " freq: " << freq << endl;
   }
-  cout << " MIN_ER: " << MIN_ER << endl;
-
+  cout << " MIN_SDR: " << MIN_SDR << endl;
 
   initNodes();
   initTree();
 
+
 }
 
-LinearSplitsTree::LinearSplitsTree(const LinearSplitsTree& ls):
-  id(ls.id), mode(ls.mode), 
-  freq(ls.freq), M(ls.M),
-  featPct(ls.featPct), SIMPLE(ls.SIMPLE), 
-  MIN_ER(ls.MIN_ER), rng(ls.rng)
+M5Tree::M5Tree(const M5Tree& m5):
+  id(m5.id), mode(m5.mode), freq(m5.freq), M(m5.M),
+  featPct(m5.featPct), SIMPLE(m5.SIMPLE), ALLOW_ALL_FEATS(m5.ALLOW_ALL_FEATS),
+  MIN_SDR(m5.MIN_SDR), rng(m5.rng)
 {
-  COPYDEBUG = ls.COPYDEBUG;
-  if (COPYDEBUG) cout << "LS copy " << id << endl;
+  COPYDEBUG = m5.COPYDEBUG;
+  if (COPYDEBUG) cout << "m5 copy " << id << endl;
   nnodes = 0;
-  nOutput = ls.nOutput;
-  nExperiences = ls.nExperiences;
-  hadError = ls.hadError;
+  nOutput = m5.nOutput;
+  nExperiences = m5.nExperiences;
+  hadError = m5.hadError;
   totalnodes = 0;
-  maxnodes = ls.maxnodes;
-  SPLIT_MARGIN = ls.SPLIT_MARGIN; 
-  LMDEBUG = ls.LMDEBUG;
-  DTDEBUG = ls.DTDEBUG;
-  SPLITDEBUG = ls.SPLITDEBUG;
-  STOCH_DEBUG = ls.STOCH_DEBUG; 
-  INCDEBUG = ls.INCDEBUG; 
-  NODEDEBUG = ls.NODEDEBUG;
+  maxnodes = m5.maxnodes;
+  SPLIT_MARGIN = m5.SPLIT_MARGIN; 
+  LMDEBUG = m5.LMDEBUG;
+  DTDEBUG = m5.DTDEBUG;
+  SPLITDEBUG = m5.SPLITDEBUG;
+  STOCH_DEBUG = m5.STOCH_DEBUG; 
+  INCDEBUG = m5.INCDEBUG; 
+  NODEDEBUG = m5.NODEDEBUG;
+  nfeat = m5.nfeat;
 
-  if (COPYDEBUG) cout << "   LS copy nodes, experiences, root, etc" << endl;
+  if (COPYDEBUG) cout << "   M5 copy nodes, experiences, root, etc" << endl;
   // copy all experiences
-  for (int i = 0; i < N_LST_EXP; i++){
-    allExp[i] = ls.allExp[i];
+  for (int i = 0; i < N_M5_EXP; i++){
+    allExp[i] = m5.allExp[i];
   }
-  if (COPYDEBUG) cout << "   LS copied exp array" << endl;
+  if (COPYDEBUG) cout << "   M5 copied exp array" << endl;
 
   // set experience pointers
-  experiences.resize(ls.experiences.size());
-  for (unsigned i = 0; i < ls.experiences.size(); i++){
+  experiences.resize(m5.experiences.size());
+  for (unsigned i = 0; i < m5.experiences.size(); i++){
     experiences[i] = &(allExp[i]);
   }
-  if (COPYDEBUG) cout << "   LS set experience pointers" << endl;
+  if (COPYDEBUG) cout << "   M5 set experience pointers" << endl;
 
   // now the tricky part, set the pointers inside the tree nodes correctly
   initNodes();
 
-  if (COPYDEBUG) cout << "   LS copy tree " << endl;
+  if (COPYDEBUG) cout << "   M5 copy tree " << endl;
   root = allocateNode();
   lastNode = root;
-  copyTree(root, ls.root);
-  if (COPYDEBUG) cout << "   LS  tree copy done" << endl;
+  copyTree(root, m5.root);
+  if (COPYDEBUG) cout << "   M5 tree copy done" << endl;
    
   if (COPYDEBUG) {
     cout << endl << "New tree: " << endl;
     printTree(root, 0);
     cout << endl;
-    cout << "  LS copy done" << endl;
+    cout << "  m5 copy done" << endl;
   }
 
 }
 
-void LinearSplitsTree::copyTree(tree_node* newNode, tree_node* origNode){
+void M5Tree::copyTree(tree_node* newNode, tree_node* origNode){
 
   int nodeId = newNode->id;
 
@@ -141,23 +148,27 @@ void LinearSplitsTree::copyTree(tree_node* newNode, tree_node* origNode){
   }
 }
 
-LinearSplitsTree* LinearSplitsTree::getCopy(){
-  LinearSplitsTree* copy = new LinearSplitsTree(*this);
+M5Tree* M5Tree::getCopy(){
+  M5Tree* copy = new M5Tree(*this);
   return copy;
 }
 
-LinearSplitsTree::~LinearSplitsTree() {
+M5Tree::~M5Tree() {
   deleteTree(root);
-  for (unsigned i = N_LST_EXP; i < experiences.size(); i++){
+  for (unsigned i = N_M5_EXP; i < experiences.size(); i++){
     delete experiences[i];
   }
   experiences.clear();
 }
 
 // here the target output will be a single value
-bool LinearSplitsTree::trainInstance(classPair &instance){
+bool M5Tree::trainInstance(classPair &instance){
 
   if (DTDEBUG) cout << id << " trainInstance" << endl;
+
+  //  featPct *= 0.99;
+
+  nfeat = instance.in.size();
 
   bool modelChanged = false;
 
@@ -165,7 +176,7 @@ bool LinearSplitsTree::trainInstance(classPair &instance){
 
   // take from static array until we run out
   tree_experience *e;
-  if (nExperiences < N_LST_EXP){
+  if (nExperiences < N_M5_EXP){
     // from statically created set of experiences
     e = &(allExp[nExperiences]);
 
@@ -209,6 +220,8 @@ bool LinearSplitsTree::trainInstance(classPair &instance){
     cout << endl << " Now have " << nExperiences << " experiences." << endl;
   }
 
+  // depending on mode/etc, maybe re-build tree
+
   // mode 0: re-build every step
   if (mode == BUILD_EVERY || nExperiences <= 1){
     rebuildTree();
@@ -228,7 +241,7 @@ bool LinearSplitsTree::trainInstance(classPair &instance){
     if (error > 0.0){
       rebuildTree();
       modelChanged = true;
-    } 
+    }
   }
 
   // mode 2: re-build every FREQ steps
@@ -241,35 +254,14 @@ bool LinearSplitsTree::trainInstance(classPair &instance){
   }
 
   if (modelChanged){
-    if (DTDEBUG || SPLITDEBUG) cout << "DT " << id << " tree re-built." << endl;
+    if (DTDEBUG) cout << "DT " << id << " tree re-built." << endl;
 
-    if (DTDEBUG || SPLITDEBUG){
+    if (DTDEBUG){
       cout << endl << "DT: " << id << endl;
       printTree(root, 0);
       cout << "Done printing tree" << endl;
     }
   }
-
-  /*
-  if (nExperiences % 100 == 0){
-    cout << endl << "DT: " << id << endl;
-    printTree(root, 0);
-    cout << "Done printing tree" << endl;
-    
-    // look at error
-    float errorSum = 0.0;
-    for (int i = 0; i < nExperiences; i++){
-      e = experiences[i];
-      std::map<float, float> answer;
-      testInstance(e->input, &answer);
-      float val = answer.begin()->first;
-      float error = fabs(val - e->output);
-      errorSum += error;
-    }
-    float avgError = errorSum / (float)nExperiences;
-    cout << "avgError: " << avgError << endl << endl;
-  }
-  */
 
   return modelChanged;
 
@@ -277,9 +269,14 @@ bool LinearSplitsTree::trainInstance(classPair &instance){
 
 
 // here the target output will be a single value
-bool LinearSplitsTree::trainInstances(std::vector<classPair> &instances){
-  if (DTDEBUG) cout << "DT trainInstances: " << instances.size() << endl;
-
+bool M5Tree::trainInstances(std::vector<classPair> &instances){
+  if (DTDEBUG) cout << id << " DT trainInstances: " 
+                            << instances.size() 
+                            << " nExp: " << nExperiences << endl;
+  nfeat = instances[0].in.size();
+  
+  //  featPct *= 0.99;
+  
   bool modelChanged = false;
 
   bool doBuild = false;
@@ -292,7 +289,7 @@ bool LinearSplitsTree::trainInstances(std::vector<classPair> &instances){
 
     // take from static array until we run out
     tree_experience *e;
-    if (nExperiences < N_LST_EXP){
+    if (nExperiences < N_M5_EXP){
       // from statically created set of experiences
       e = &(allExp[nExperiences]);
 
@@ -328,27 +325,6 @@ bool LinearSplitsTree::trainInstances(std::vector<classPair> &instances){
       }
       cout << endl << " Now have " << nExperiences << " experiences." << endl;
     }
-
-    /*
-    if (nExperiences % 100 == 0){
-      cout << endl << "DT: " << id << endl;
-      printTree(root, 0);
-      cout << "Done printing tree" << endl;
-      
-      // look at error
-      float errorSum = 0.0;
-      for (int i = 0; i < nExperiences; i++){
-        e = experiences[i];
-        std::map<float, float> answer;
-        testInstance(e->input, &answer);
-        float val = answer.begin()->first;
-        float error = fabs(val - e->output);
-        errorSum += error;
-      }
-      float avgError = errorSum / (float)nExperiences;
-      cout << "avgError: " << avgError << endl << endl;
-    }
-    */
 
     // depending on mode/etc, maybe re-build tree
 
@@ -394,9 +370,9 @@ bool LinearSplitsTree::trainInstances(std::vector<classPair> &instances){
   }
 
   if (modelChanged){
-    if (DTDEBUG || SPLITDEBUG) cout << "DT " << id << " tree re-built." << endl;
+    if (DTDEBUG) cout << "DT " << id << " tree re-built." << endl;
 
-    if (DTDEBUG || SPLITDEBUG){
+    if (DTDEBUG){
       cout << endl << "DT: " << id << endl;
       printTree(root, 0);
       cout << "Done printing tree" << endl;
@@ -408,21 +384,17 @@ bool LinearSplitsTree::trainInstances(std::vector<classPair> &instances){
 }
 
 
-void LinearSplitsTree::rebuildTree(){
+void M5Tree::rebuildTree(){
   //cout << "rebuild tree " << id << " on exp: " << nExperiences << endl;
-  //deleteTree(root);
-
-  // re-calculate avg error for root
-  root->avgError = calcAvgErrorforSet(experiences);
-
+  //  deleteTree(root);
   buildTree(root, experiences, false);
   //cout << "tree " << id << " rebuilt. " << endl;
 }
 
 
 // TODO: here we want to return the probability of the output value being each of the possible values, in the stochastic case
-void LinearSplitsTree::testInstance(const std::vector<float> &input, std::map<float, float>* retval){
-  if (DTDEBUG) cout << "testInstance on tree " << id << endl;
+void M5Tree::testInstance(const std::vector<float> &input, std::map<float, float>* retval){
+  if (DTDEBUG) cout << "testInstance" << endl;
 
   retval->clear();
 
@@ -441,7 +413,7 @@ void LinearSplitsTree::testInstance(const std::vector<float> &input, std::map<fl
 
 }
 
-float LinearSplitsTree::getConf(const std::vector<float> &input){
+float M5Tree::getConf(const std::vector<float> &input){
   if (DTDEBUG) cout << "numVisits" << endl;
 
   // in case the tree is empty
@@ -449,8 +421,9 @@ float LinearSplitsTree::getConf(const std::vector<float> &input){
     return 0;
   }
 
-  if (lastNode == NULL)
+  if (lastNode == NULL){
     return 0;
+  }
 
   // follow through tree to leaf
   //tree_node* leaf = traverseTree(root, input);
@@ -468,9 +441,8 @@ float LinearSplitsTree::getConf(const std::vector<float> &input){
 // to get more info on potential splits
 
 
-
 // init the tree
-void LinearSplitsTree::initTree(){
+void M5Tree::initTree(){
   if (DTDEBUG) cout << "initTree()" << endl;
   root = allocateNode();
 
@@ -486,7 +458,7 @@ void LinearSplitsTree::initTree(){
 
 
 // init a tree node
-void LinearSplitsTree::initTreeNode(tree_node* node){
+void M5Tree::initTreeNode(tree_node* node){
   if (DTDEBUG) cout << "initTreeNode()";
 
   node->id = nnodes++;
@@ -495,7 +467,7 @@ void LinearSplitsTree::initTreeNode(tree_node* node){
   totalnodes++;
   if (totalnodes > maxnodes){
     maxnodes = totalnodes;
-    if (DTDEBUG) cout << id << " LS MAX nodes: " << maxnodes << endl;
+    if (DTDEBUG) cout << id << " M5 MAX nodes: " << maxnodes << endl;
   }
 
   // split criterion
@@ -504,9 +476,7 @@ void LinearSplitsTree::initTreeNode(tree_node* node){
 
   // current data
   node->nInstances = 0;
-
-  // coefficients
-  node->constant = 0.0;
+  node->constant = 0;
 
   // coefficients will get resized later
   //  node->coefficients.resize(2,0);
@@ -516,12 +486,10 @@ void LinearSplitsTree::initTreeNode(tree_node* node){
   node->r = NULL;
 
   node->leaf = true;
-  node->avgError = 10000;
 
 }
 
-/** delete current tree */
-void LinearSplitsTree::deleteTree(tree_node* node){
+void M5Tree::deleteTree(tree_node* node){
   if (DTDEBUG) cout << "deleteTree, node=" << node->id << endl;
 
   if (node==NULL)
@@ -531,7 +499,7 @@ void LinearSplitsTree::deleteTree(tree_node* node){
 
   node->nInstances = 0;
   node->coefficients.clear();
- 
+
   //recursively call deleteTree on children
   // then delete them
   if (!node->leaf){
@@ -550,17 +518,15 @@ void LinearSplitsTree::deleteTree(tree_node* node){
     }
   }
 
+  node->leaf  = true;
   node->dim = -1;
-  node->leaf = true;
-  node->avgError = 10000;
   node->val = -1;
   node->constant = 0;
 }
 
 
-/** Get the correct child of this node based on the input */
-LinearSplitsTree::tree_node* LinearSplitsTree::getCorrectChild(tree_node* node,
-                                                               const std::vector<float> &input){
+M5Tree::tree_node* M5Tree::getCorrectChild(tree_node* node,
+                                           const std::vector<float> &input){
 
   if (DTDEBUG) cout << "getCorrectChild, node=" << node->id << endl;
 
@@ -571,9 +537,8 @@ LinearSplitsTree::tree_node* LinearSplitsTree::getCorrectChild(tree_node* node,
 
 }
 
-/** Traverse the tree to the leaf for this input. */
-LinearSplitsTree::tree_node* LinearSplitsTree::traverseTree(tree_node* node,
-                                                            const std::vector<float> &input){
+M5Tree::tree_node* M5Tree::traverseTree(tree_node* node,
+                                        const std::vector<float> &input){
 
   if (DTDEBUG) cout << "traverseTree, node=" << node->id << endl;
 
@@ -585,12 +550,10 @@ LinearSplitsTree::tree_node* LinearSplitsTree::traverseTree(tree_node* node,
 }
 
 
-/** Decide if this passes the test */
-bool LinearSplitsTree::passTest(int dim, float val, const std::vector<float> &input){
+bool M5Tree::passTest(int dim, float val, const std::vector<float> &input){
   if (DTDEBUG) cout << "passTest, dim=" << dim << ",val=" << val 
                     << ",input["<<dim<<"]=" << input[dim] <<endl;
 
-  
   if (input[dim] > val)
     return false;
   else
@@ -599,8 +562,7 @@ bool LinearSplitsTree::passTest(int dim, float val, const std::vector<float> &in
 }
 
 
-/** Build the tree from this node down using this set of experiences. */
-void LinearSplitsTree::buildTree(tree_node *node,
+void M5Tree::buildTree(tree_node *node,
                        const std::vector<tree_experience*> &instances,
                        bool changed){
   if(DTDEBUG) cout << "buildTree, node=" << node->id
@@ -619,7 +581,6 @@ void LinearSplitsTree::buildTree(tree_node *node,
   // first, add instances to tree
   node->nInstances = instances.size();
 
-  // add each output to this node
   bool allSame = true;
   float val0 = instances[0]->output;
   for (unsigned i = 1; i < instances.size(); i++){
@@ -631,51 +592,79 @@ void LinearSplitsTree::buildTree(tree_node *node,
 
   // see if they're all the same
   if (allSame){
-    makeLeaf(node, instances);
+    makeLeaf(node);
+    node->constant = instances[0]->output;
     if (DTDEBUG){
       cout << "Tree " << id << " node " << node->id 
            << " All " << node->nInstances
            << " classified with output "
-           << instances[0]->output << ", " << node->constant << endl;
+           << instances[0]->output << endl;
     }
     return;
   }
 
-  // check if linear model has no error
-  if (node != root && node->avgError < MIN_ER){
-    makeLeaf(node, instances);
-    if (node->avgError < MIN_ER){
-      if (DTDEBUG) {
-      cout << "Tree " << id << " node " << node->id
-           << " has low error " << node->avgError << " keeping as leaf" << endl;
-      }
+  // check if this is a leaf and linear model has no error
+  if (node->leaf && node->coefficients.size() > 0) {
+    //cout << "Node " << node->id << " is leaf, checking lm" << endl;
+    float errorSum = 0;
+    for (unsigned i = 0; i < instances.size(); i++){
+      // get prediction for instance and compare with actual output
+      tree_node* leaf = traverseTree(node, instances[i]->input);
+      std::map<float, float> retval;
+      leafPrediction(leaf, instances[i]->input, &retval);
+      float prediction = retval.begin()->first;
+      float absError = fabs(prediction - instances[i]->output);
+      errorSum += absError;
+      //cout << "instance " << i << " leaf predicted: " << prediction
+      //     << " actual: " << instances[i]->output
+      //   << " error: " << absError << endl;
+    }
+    float avgError = errorSum / (float)instances.size();
+    if (avgError < 0.001){
+      //      cout << "stick with linear model" << endl;
       return;
     }
   }
-  
-  // if not, calculate ER to determine best split
-  if (SPLITDEBUG) cout << endl << "Creating new decision node " << id << "-" << node->id << endl;
+
+  // if not, calculate SDR to determine best split
+  if (SPLITDEBUG) cout << endl << "Creating new decision node" << endl;
 
   node->leaf = false;
   //node->nInstances++;
 
-  float bestER = -1.0;
+  float bestSDR = -1.0;
   int bestDim = -1;
   float bestVal = -1;
   std::vector<tree_experience*> bestLeft;
   std::vector<tree_experience*> bestRight;
-  float leftError = 10000;
-  float rightError = 10000;
 
-  testPossibleSplits(node->avgError, instances, &bestER, &bestDim, &bestVal, &bestLeft, &bestRight, &leftError, &rightError);
+  testPossibleSplits(instances, &bestSDR, &bestDim, &bestVal, &bestLeft, &bestRight);
 
-  implementSplit(node, instances, bestER, bestDim, bestVal, bestLeft, bestRight, changed, leftError, rightError);
+  implementSplit(node, instances, bestSDR, bestDim, bestVal, bestLeft, bestRight, changed);
+
+  // possibly replace split node with linear regression model
+  pruneTree(node, instances);
 
 }
 
 
-void LinearSplitsTree::makeLeaf(tree_node* node, const std::vector<tree_experience*> &instances){
+void M5Tree::makeLeaf(tree_node* node){
 
+  removeChildren(node);
+
+  // make sure there are enough coefficients for all the features
+  // and that they are 0
+  if (node->coefficients.size() != (unsigned)nfeat){
+    node->coefficients.resize(nfeat, 0);
+  }
+  for (unsigned i = 0; i < node->coefficients.size(); i++){
+    node->coefficients[i] = 0;
+  }
+
+}
+
+
+void M5Tree::removeChildren(tree_node* node){
   // check on children
   if (node->l != NULL){
     deleteTree(node->l);
@@ -690,53 +679,150 @@ void LinearSplitsTree::makeLeaf(tree_node* node, const std::vector<tree_experien
   }
 
   node->leaf = true;
+}
 
-  // fit linear model
+
+void M5Tree::pruneTree(tree_node *node,
+                       const std::vector<tree_experience*> &instances){
+  if (LMDEBUG || DTDEBUG){
+    printTree(root, 0);
+    cout << "pruneTree, node=" << node->id
+         << ",nInstances:" << instances.size() << endl;
+  }
+
+  // TODO: no pruning right now
+  //  return;
+
+  // calculate error of current subtree
+  float subtreeErrorSum = 0;
+  for (unsigned i = 0; i < instances.size(); i++){
+    
+    // get prediction for instance and compare with actual output
+    tree_node* leaf = traverseTree(node, instances[i]->input);
+    std::map<float, float> retval;
+    leafPrediction(leaf, instances[i]->input, &retval);
+    float prediction = retval.begin()->first;
+    float absError = fabs(prediction - instances[i]->output);
+    subtreeErrorSum += absError;
+    if (LMDEBUG || DTDEBUG){
+      cout << "instance " << i << " subtree predicted: " << prediction
+           << " actual: " << instances[i]->output
+           << " error: " << absError << endl;
+    }
+  }
+  if (instances.size() < 3){
+    if (LMDEBUG || DTDEBUG) cout << "instances size <= 2!!!" << endl;
+    return;
+  }
+
+  float avgTreeError = subtreeErrorSum / (float)instances.size();
+
+  // if this is zero error, we're not going to replace it
+  if (false && avgTreeError <= 0.0001){
+    if (LMDEBUG || DTDEBUG) 
+      cout << "Sub-tree is perfect (" << avgTreeError
+           << "), do not attempt LM" << endl;
+    return;
+  }
+                                                                
+  // figure out tree feats used
+  std::vector<bool> treeFeatsUsed(instances[0]->input.size(), false);
+  getFeatsUsed(node, &treeFeatsUsed);
+  int nTreeFeatsUsed = 0;
+  for (unsigned i = 0; i < treeFeatsUsed.size(); i++){
+    if (treeFeatsUsed[i])
+      nTreeFeatsUsed++;
+  }
+
+  // Just use them all... otherwise we ignore some that weren't good enough
+  // for splitting, but are good here
+  std::vector<bool> featsUsed(instances[0]->input.size(), true);
+  int nFeatsUsed = featsUsed.size();
+
+  // or just use ones allowed by subtree
+  if (!ALLOW_ALL_FEATS){
+    featsUsed = treeFeatsUsed;
+    nFeatsUsed = nTreeFeatsUsed;
+  }
+
+  // no features to build linear model on
+  if (nFeatsUsed == 0){
+    if (LMDEBUG || DTDEBUG) cout << "No features for LM" << endl;
+    return;
+  }
+
+  // add on error bonus based on nInstances and nParams
+  float treeErrorEst = avgTreeError;
+  float denom = (instances.size() - nTreeFeatsUsed);
+  if (denom < 1){
+    denom = 0.5;
+    if (LMDEBUG) {
+      cout << "denom of tree error factor is " << denom 
+           << " with nInst " << instances.size() 
+           << " nfeats: " << nTreeFeatsUsed << endl;
+    }
+  }
+  treeErrorEst *= (instances.size() + nTreeFeatsUsed) / denom;
+
+  // fit linear model to this set of instances
+  float lmErrorSum = 0;
+  int nlmFeats = 0;
   if (SIMPLE)
-    node->avgError = fitSimpleLinearModel(instances, &(node->constant), &(node->coefficients));
+    nlmFeats = fitSimpleLinearModel(node, instances, featsUsed, nFeatsUsed, &lmErrorSum);
   else 
-    node->avgError = fitMultiLinearModel(instances, &(node->constant), &(node->coefficients));
+    nlmFeats = fitLinearModel(node, instances, featsUsed, nFeatsUsed, &lmErrorSum);
 
-  if (DTDEBUG || LMDEBUG){
-    cout << "make leaf, fit linear model with constant " << node->constant 
-         << "  error: " << node->avgError << endl;
+  float avgLMError = lmErrorSum / (float)instances.size();
+
+  float lmErrorEst = avgLMError;
+  float denom2 = (instances.size() - nlmFeats);
+  if (denom2 < 1){
+    denom2 = 0.5;
+    if (LMDEBUG) {
+      cout << "denom2 of lm error factor is " << denom2
+           << " with nInst " << instances.size() 
+           << " nfeats: " << nlmFeats << endl;
+    }
+  }
+  lmErrorEst *= (instances.size() + nlmFeats) / denom2; 
+
+  // replace subtree with linear model?
+  if (LMDEBUG || DTDEBUG) {
+    cout << "Sub-Tree Error: " << treeErrorEst << ", lm Error: "
+         << lmErrorEst << endl;
+  }
+  //  if (lmErrorEst < (treeErrorEst + 0.0001)){
+  if (lmErrorEst < (treeErrorEst + 0.1*MIN_SDR)){
+  //if (lmErrorEst < treeErrorEst){
+    if (LMDEBUG || DTDEBUG)
+      cout << node->id << " replace tree with linear model" << endl;
+    removeChildren(node);
+  } else {
+    // remove coefficients again, for memory
+    //    node->coefficients.clear();
   }
 
 }
 
 
-float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> &instances,
-                                            float *bestConstant, 
-                                            std::vector<float> *bestCoefficients){
-  if(DTDEBUG || LMDEBUG) cout << "fitMultiLinearModel"
+int M5Tree::fitLinearModel(tree_node *node,
+                           const std::vector<tree_experience*> &instances,
+                           std::vector<bool> featureMask,
+                           int nFeats, float* resSum){
+  
+  if(DTDEBUG || LMDEBUG) cout << "fitLinearModel, node=" << node->id
                               << ",nInstances:" << instances.size() << endl;
 
   // make sure there are enough coefficients for all the features
-  if (bestCoefficients->size() != instances[0]->input.size()){
-    bestCoefficients->resize(instances[0]->input.size(), 0);  
+  if (node->coefficients.size() != instances[0]->input.size()){
+    node->coefficients.resize(instances[0]->input.size(), 0);
   }
 
-  // in case of size 1
-  if (instances.size() == 1){
-    *bestConstant = instances[0]->output;
-    for (unsigned i = 0; i < bestCoefficients->size(); i++){
-      (*bestCoefficients)[i] = 0.0;
-    }
-    if (LMDEBUG || SPLITDEBUG){
-      cout << "  Singleton constant: " 
-           << *bestConstant << " avgE: " << 0 << endl;
-    }
-    return 0;
-  }
-
-  // loop through all features, try simple single variable regression
-  // keep track of error/coefficient of each
-  float bestError = 100000;
-  float avgError = 100000;
+  node->constant = 0.0;
   bool doRegression = true;
   int ntimes = 0;
-
-  std::vector<bool> featureMask(bestCoefficients->size(), true);
+  int nlmFeats = 10;
+  (*resSum) = 1000000;
 
   while (doRegression){
     //cout << id << " Attempt linear model " << ntimes << endl;
@@ -744,12 +830,7 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
 
     int nObs = (int)instances.size();
 
-    int nFeats = 0;
-    for (unsigned i = 0; i < featureMask.size(); i++){
-      if (featureMask[i]) nFeats++;
-    }
-    if (nFeats < 1)
-      break;
+    //cout << "with nObs: " << nObs << " and nFeats: " << nFeats << endl;
 
     // no feats or obs, no model to build
     if (nObs == 0 || nFeats == 0)
@@ -771,13 +852,24 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
       // go through all features
       int featIndex = 1;
       for (unsigned j = 0; j < featureMask.size(); j++){
-        (*bestCoefficients)[j] = 0;
+        node->coefficients[j] = 0;
         if (!featureMask[j])
           continue;
-
+      
         if (constants[j] && e->input[j] != instances[0]->input[j]){
           constants[j] = false;
         }
+
+	/*
+        if (rng.uniform() < featPct){
+          featureMask[j] = false;
+          nFeats--;
+          if (nFeats > 0)
+            continue;
+          else
+            break;
+        }
+	*/
 
         if (i == nObs-1 && constants[j]){
           //cout << "PROBLEM: feat " << j << " is constant!" << endl;
@@ -804,7 +896,7 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
       if (LMDEBUG) cout << " out: " << e->output << endl;
     }
 
-    if (foundProblem)
+    if (foundProblem || nFeats == 0)
       continue;
 
     // make vector of 1s
@@ -822,13 +914,13 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
     Real mval = Sum(Y) / nObs;
     YC = Y - Ones * mval;
 
-    // form sum of squares and product matrix
-    //    [use << rather than = for copying Matrix into SymmetricMatrix]
-    SymmetricMatrix SSQ;
-    SSQ << XC.t() * XC;
-
     Try {
 
+      // form sum of squares and product matrix
+      //    [use << rather than = for copying Matrix into SymmetricMatrix]
+      SymmetricMatrix SSQ;
+      SSQ << XC.t() * XC;
+      
       ///////////////////////////
       // Cholesky Method
       LowerTriangularMatrix L = Cholesky(SSQ);
@@ -839,13 +931,12 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
       
       //////////////////////////
       // Least Squares Method
-      /*
       // calculate estimate
       //    [bracket last two terms to force this multiplication first]
       //    [ .i() means inverse, but inverse is not explicity calculated]
-      ColumnVector A = SSQ.i() * (XC.t() * YC);
+      //ColumnVector A = SSQ.i() * (XC.t() * YC);
       //////////////////////////
-      */
+  
   
       // calculate estimate of constant term
       //    [AsScalar converts 1x1 matrix to Real]
@@ -859,24 +950,28 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
 
       // print out answers
       // for each instance
-      bestError = 0;
+      (*resSum) = 0;
       for (int i = 0; i < nObs; i++){
         if (DTDEBUG || LMDEBUG){
           cout << "instance " << i << " linear model predicted: " << Fitted(i+1)
                << " actual: " << instances[i]->output
                << " error: " << Residual(i+1) << endl;
         }
-        bestError += fabs(Residual(i+1));
+        (*resSum) += fabs(Residual(i+1));
       }
-      avgError = bestError / (float)nObs;
+
 
       // coeff
+      nlmFeats = 0;
       for (int i = 0; i < nFeats; i++){
         if (DTDEBUG || LMDEBUG) cout << "Coeff " << i << " on feat: " << featIndices[i] << " is " << A(i+1) << endl;
-        (*bestCoefficients)[featIndices[i]] = A(i+1);
+        node->coefficients[featIndices[i]] = A(i+1);
+        if (A(i+1) != 0)
+          nlmFeats++;
       }
       if (DTDEBUG || LMDEBUG) cout << "constant is " << a << endl;
-      *bestConstant = a;
+      node->constant = a;
+
 
     }
 
@@ -886,31 +981,8 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
       //cout << ntimes << " linear regression had exception" << endl;
       //<< BaseException::what() <<endl;
 
-      /*
-      for (int i = 0; i < nObs; i++){
-        tree_experience *e = instances[i];
-        cout << "Obs: " << i;
-        
-        // go through all features
-        int featIndex = 1;
-        for (unsigned j = 0; j < featureMask.size(); j++){
-          node->coefficients[j] = 0;
-          if (!featureMask[j])
-            continue;
-          
-          
-          cout << " Feat " << featIndex << " index " << j
-               << " val " << X(i+1,featIndex) << ",";
-          
-          featIndex++;
-        }
-        
-        cout << " out: " << e->output << endl;
-      }
-      */
-
       // tried this already, stop now
-      if (ntimes > 2 || nFeats < 2){
+      if (ntimes > 1 || nFeats < 2){
         //cout << "max regression" << endl;
         doRegression = false;
         break;
@@ -924,50 +996,36 @@ float LinearSplitsTree::fitMultiLinearModel(const std::vector<tree_experience*> 
         }
       }
       continue;
-    } // catch
+    }
 
     // it worked, dont need to do it again
     doRegression = false;
 
   }
 
-  // return error
-  return avgError;
+  // return # features used
+  return nlmFeats;
 
 }
 
-
-float LinearSplitsTree::fitSimpleLinearModel(const std::vector<tree_experience*> &instances,
-                                             float *bestConstant, std::vector<float> *bestCoefficients){
-  if(DTDEBUG || LMDEBUG || SPLITDEBUG) cout << " fitSimpleLinearModel, "
+int M5Tree::fitSimpleLinearModel(tree_node *node,
+                                 const std::vector<tree_experience*> &instances,
+                                 std::vector<bool> featureMask,
+                                 int nFeats, float* resSum){
+  if(DTDEBUG || LMDEBUG) cout << "fitSimpleLinearModel, node=" << node->id
                               << ",nInstances:" << instances.size() << endl;
   
   // make sure there are enough coefficients for all the features
-  if (bestCoefficients->size() != instances[0]->input.size()){
-    bestCoefficients->resize(instances[0]->input.size(), 0);  
+  if (node->coefficients.size() != instances[0]->input.size()){
+    node->coefficients.resize(instances[0]->input.size(), 0);
   }
 
   // loop through all features, try simple single variable regression
   // keep track of error/coefficient of each
   int bestFeat = -1;
   float bestCoeff = -1;
-  float bestError = 100000;
-  *bestConstant = -1;
-
-  // in case of size 1
-  if (instances.size() == 1){
-    bestFeat = 0;
-    *bestConstant = instances[0]->output;
-    for (unsigned i = 0; i < bestCoefficients->size(); i++){
-      (*bestCoefficients)[i] = 0.0;
-    }
-    bestError = 0;
-    if (LMDEBUG || SPLITDEBUG){
-      cout << "  Singleton constant: " 
-           << *bestConstant << " avgE: " << bestError << endl;
-    }
-    return 0;
-  }
+  float bestError = 1000000;
+  float bestConstant = -1;
 
   std::vector<float> xsum(instances[0]->input.size(), 0);
   std::vector<float> xysum(instances[0]->input.size(), 0);
@@ -981,6 +1039,7 @@ float LinearSplitsTree::fitSimpleLinearModel(const std::vector<tree_experience*>
     
     // go through all features
     for (unsigned j = 0; j < instances[0]->input.size(); j++){
+      if (!featureMask[j]) continue;
       if (LMDEBUG) cout << ", F" << j << ": " << e->input[j];
       xsum[j] += e->input[j];
       xysum[j] += (e->input[j] * e->output);
@@ -989,23 +1048,21 @@ float LinearSplitsTree::fitSimpleLinearModel(const std::vector<tree_experience*>
     ysum += e->output;
     if (LMDEBUG) cout << ", out: " << e->output << endl;
   }
-
+  
   // now go through all features and calc coeff and constant
   for (unsigned j = 0; j < instances[0]->input.size(); j++){
+    if (!featureMask[j]) continue;
+    /*
+      if (rng.uniform() < featPct){
+      continue;
+      }
+    */
     float coeff = (xysum[j] - xsum[j]*ysum/nObs)/(x2sum[j]-(xsum[j]*xsum[j])/nObs);
     float constant = (ysum/nObs) - coeff*(xsum[j]/nObs);
-
-    // so we don't get absurd coefficients that get rounded off earlier
-    if (fabs(coeff) < 1e-5){
-      coeff = 0.0;
-    }
-    if (fabs(constant) < 1e-5){
-      constant = 0.0;
-    }
-
+    
     if (LMDEBUG) {
       cout << "Feat " << j << " coeff: " << coeff << ", constant: " 
-           << constant  << endl;
+           << constant << " mask: " << featureMask[j] << endl;
     }
 
     // now try to make predictions and see what error is
@@ -1017,85 +1074,92 @@ float LinearSplitsTree::fitSimpleLinearModel(const std::vector<tree_experience*>
       if (LMDEBUG) cout << "Instance " << i << " error: " << error << endl;
       errorSum += error;
     }
-    float avgError = errorSum / (float)nObs;
-    if (LMDEBUG) cout << "avgError: " << avgError << endl;
+    if (LMDEBUG) cout << "eSum: " << errorSum << endl;
 
     // check if this is the best
-    if (avgError < bestError){
-      bestError = avgError;
+    if (errorSum < bestError){
+      bestError = errorSum;
       bestFeat = j;
-      *bestConstant = constant;
+      bestConstant = constant;
       bestCoeff = coeff;
     }
   }
   
-  if (LMDEBUG || SPLITDEBUG){
-    cout << "  LST feat: " << bestFeat << " coeff: " 
+  if (LMDEBUG){
+    cout << "SLM feat: " << bestFeat << " coeff: " 
          << bestCoeff << " constant: " 
-         << *bestConstant << " avgE: " << bestError << endl;
+         << bestConstant << " avgE: " << (bestError/nObs) << endl;
   }
 
   if (bestFeat < 0 || bestFeat > (int)instances[0]->input.size()){
-    for (unsigned i = 0; i < bestCoefficients->size(); i++){
-      (*bestCoefficients)[i] = 0.0;
+    node->constant = 0.0;
+    for (unsigned i = 0; i < node->coefficients.size(); i++){
+      node->coefficients[i] = 0.0;
     }
-    *bestConstant = 0.0;
-    return 100000;
+    (*resSum) = 1000000;
+    return 10;
   }
 
-  // fill in coeff vector
-  for (unsigned i = 0; i < bestCoefficients->size(); i++){
-    if (i == (unsigned)bestFeat)
-      (*bestCoefficients)[i] = bestCoeff;
-    else
-      (*bestCoefficients)[i] = 0.0; 
+  // pick best feature somehow
+  node->constant = bestConstant;
+  for (unsigned i = 0; i < node->coefficients.size(); i++){
+    if (i == (unsigned)bestFeat){
+      node->coefficients[i] = bestCoeff;
+      if (LMDEBUG) cout << "Set coefficient on feat " << i << " to " << bestCoeff << endl;
+    }
+    else {
+      node->coefficients[i] = 0.0;
+    }
   }
 
-  return bestError;
+  (*resSum) = bestError;
+
+  // only use 1 input feature this way
+  return 1;
  
 }
 
 
 
 
-void LinearSplitsTree::implementSplit(tree_node* node, 
-                                      const std::vector<tree_experience*> &instances,
-                                      float bestER, int bestDim,
-                                      float bestVal, 
-                                      const std::vector<tree_experience*> &bestLeft,
-                                      const std::vector<tree_experience*> &bestRight,
-                                      bool changed, float leftError, float rightError){
+void M5Tree::implementSplit(tree_node* node, 
+                            const std::vector<tree_experience*> &instances,
+                            float bestSDR, int bestDim,
+                            float bestVal, 
+                            const std::vector<tree_experience*> &bestLeft,
+                            const std::vector<tree_experience*> &bestRight,
+                            bool changed){
   if (DTDEBUG) cout << "implementSplit node=" << node->id
-                    << ",er=" << bestER
+                    << ",sdr=" << bestSDR
                     << ",dim=" << bestDim
                     << ",val=" << bestVal 
                     << ",chg=" << changed << endl;
 
 
   // see if this should still be a leaf node
-  if (bestER < MIN_ER){
-    makeLeaf(node, instances);
-
+  if (bestSDR < MIN_SDR){
+    makeLeaf(node);
+    float valSum = 0;
+    for (unsigned i = 0; i < instances.size(); i++){
+      valSum += instances[i]->output;
+    }
+    float avg = valSum / (float)(instances.size());
+    node->constant = avg;
     if (SPLITDEBUG || STOCH_DEBUG){
-      cout << " DT " << id << " Node " << node->id << " Poor er "
+      cout << "DT " << id << " Node " << node->id << " Poor sdr "
            << node->nInstances
            << " instances classified at leaf " << node->id
-           << " with er " << bestER << " constant: " << node->constant << endl;
+           << " with multiple outputs " << endl;
     }
     return;
   }
 
-  //cout << id << " implement split with er " << bestER << endl;
-
   // see if this split changed or not
   // assuming no changes above
   if (!changed && node->dim == bestDim && node->val == bestVal
-      && !node->leaf && node->l != NULL && node->r != NULL){
+       && !node->leaf && node->l != NULL && node->r != NULL){
     // same split as before.
-    if (DTDEBUG || SPLITDEBUG) cout << "Same split as before " << node->id << endl;
-    node->l->avgError = leftError;
-    node->r->avgError = rightError;
-
+    if (DTDEBUG || SPLITDEBUG) cout << "Same split as before" << endl;
 
     // see which leaf changed
     if (bestLeft.size() > (unsigned)node->l->nInstances){
@@ -1121,7 +1185,7 @@ void LinearSplitsTree::implementSplit(tree_node* node,
 
   if (SPLITDEBUG) cout << "Best split was cut with val " << node->val
                        << " on dim " << node->dim
-                       << " with er: " << bestER << endl;
+                       << " with sdr: " << bestSDR << endl;
 
   if (DTDEBUG) cout << "Left has " << bestLeft.size()
                     << ", right has " << bestRight.size() << endl;
@@ -1132,7 +1196,7 @@ void LinearSplitsTree::implementSplit(tree_node* node,
          << " right: " << bestRight.size() << endl;
     cout << "Split was cut with val " << node->val
          << " on dim " << node->dim
-         << " with er: " << bestER << endl;
+         << " with sdr: " << bestSDR << endl;
     exit(-1);
   }
 
@@ -1149,22 +1213,50 @@ void LinearSplitsTree::implementSplit(tree_node* node,
 
   // recursively build the sub-trees to this one
   if (DTDEBUG) cout << "Building left tree for node " << node->id << endl;
-  node->l->avgError = leftError;
   buildTree(node->l, bestLeft, true);
   if (DTDEBUG) cout << "Building right tree for node " << node->id << endl;
-  node->r->avgError = rightError;
   buildTree(node->r, bestRight, true);
 
 }
 
 
-void LinearSplitsTree::testPossibleSplits(float avgError, const std::vector<tree_experience*> &instances,
-                                          float *bestER, int *bestDim,
-                                          float *bestVal, 
-                                          std::vector<tree_experience*> *bestLeft,
-                                          std::vector<tree_experience*> *bestRight,
-                                          float *bestLeftError, float *bestRightError) {
-  if (DTDEBUG || SPLITDEBUG) cout << "testPossibleSplits, error=" << avgError << endl;
+void M5Tree::getFeatsUsed(tree_node* node, std::vector<bool> *featsUsed){
+
+  // if leaf, ones from linear model
+  if (node->leaf){
+    //cout << "coeff size: " << node->coefficients.size() << endl;
+    for (unsigned i = 0; i < node->coefficients.size(); i++){
+      if (node->coefficients[i] != 0){
+        if (LMDEBUG || DTDEBUG) cout << "Leaf node, used coeff " << i << endl;
+        (*featsUsed)[i] = true;
+      }
+    }
+    return;
+  }
+
+  // otherwise see what split was used
+  // and call for left and right sub-trees
+  (*featsUsed)[node->dim] = true;
+  if (LMDEBUG || DTDEBUG) cout << "Split node, used feat " << node->dim << endl;
+
+  getFeatsUsed(node->l, featsUsed);
+  getFeatsUsed(node->r, featsUsed);
+
+  return;
+}
+
+
+void M5Tree::testPossibleSplits(const std::vector<tree_experience*> &instances,
+                                float *bestSDR, int *bestDim,
+                                float *bestVal, 
+                                std::vector<tree_experience*> *bestLeft,
+                                std::vector<tree_experience*> *bestRight) {
+  if (DTDEBUG) cout << "testPossibleSplits" << endl;
+
+
+  // calculate sd for the set
+  float sd = calcSDforSet(instances);
+  //if (DTDEBUG) cout << "I: " << I << endl;
 
   int nties = 0;
 
@@ -1185,25 +1277,21 @@ void LinearSplitsTree::testPossibleSplits(float avgError, const std::vector<tree
       // here (decision is taken from the random set that are left)
       if (rng.uniform() < featPct)
         continue;
-      
+
       std::vector<tree_experience*> left;
       std::vector<tree_experience*> right;
-      float leftError = 10000;
-      float rightError = 10000;
 
       // splits that are cuts
       float splitval = (*j);
-      float er = calcER(idim, splitval, instances, avgError, left, right, &leftError, &rightError);
+      float sdr = calcSDR(idim, splitval, instances, sd, left, right);
 
-      if (SPLITDEBUG){
-        cout << id << " CUT split val " << splitval
-             << " on dim: " << idim << " had er "
-             << er << endl;
-      }
+      if (SPLITDEBUG) cout << " CUT split val " << splitval
+                           << " on dim: " << idim << " had sdr "
+                           << sdr << endl;
 
-      // see if this is the new best er
-      compareSplits(er, idim, splitval, left, right, &nties, leftError, rightError,
-                    bestER, bestDim, bestVal,bestLeft, bestRight, bestLeftError, bestRightError);
+      // see if this is the new best sdr
+      compareSplits(sdr, idim, splitval, left, right, &nties,
+                    bestSDR, bestDim, bestVal, bestLeft, bestRight);
 
 
     } // j loop
@@ -1212,24 +1300,21 @@ void LinearSplitsTree::testPossibleSplits(float avgError, const std::vector<tree
 
 
 
-/** Decide if this split is better. */
-void LinearSplitsTree::compareSplits(float er, int dim, float val, 
-                                     const std::vector<tree_experience*> &left, 
-                                     const std::vector<tree_experience*> &right,
-                                     int *nties, float leftError, float rightError,
-                                     float *bestER, int *bestDim,
-                                     float *bestVal, 
-                                     std::vector<tree_experience*> *bestLeft,
-                                     std::vector<tree_experience*> *bestRight,
-                                     float *bestLeftError, float *bestRightError){
-  if (DTDEBUG) cout << "compareSplits er=" << er << ",dim=" << dim
+void M5Tree::compareSplits(float sdr, int dim, float val, 
+                           const std::vector<tree_experience*> &left, 
+                           const std::vector<tree_experience*> &right,
+                           int *nties, float *bestSDR, int *bestDim,
+                           float *bestVal, 
+                           std::vector<tree_experience*> *bestLeft,
+                           std::vector<tree_experience*> *bestRight){
+  if (DTDEBUG) cout << "compareSplits sdr=" << sdr << ",dim=" << dim
                     << ",val=" << val  <<endl;
 
 
   bool newBest = false;
 
   // if its a virtual tie, break it randomly
-  if (fabs(*bestER - er) < SPLIT_MARGIN){
+  if (fabs(*bestSDR - sdr) < SPLIT_MARGIN){
     //cout << "Split tie, making random decision" << endl;
 
     (*nties)++;
@@ -1247,7 +1332,7 @@ void LinearSplitsTree::compareSplits(float er, int dim, float val,
   }
 
   // if its clearly better, set this as the best split
-  else if (er > *bestER){
+  else if (sdr > *bestSDR){
     newBest = true;
     *nties = 1;
   }
@@ -1255,31 +1340,27 @@ void LinearSplitsTree::compareSplits(float er, int dim, float val,
 
   // set the split features
   if (newBest){
-    *bestER = er;
+    *bestSDR = sdr;
     *bestDim = dim;
     *bestVal = val;
     *bestLeft = left;
     *bestRight = right;
-    *bestLeftError = leftError;
-    *bestRightError = rightError;
     if (SPLITDEBUG){
-      cout << "  New best er: " << *bestER
+      cout << "  New best sdr: " << *bestSDR
            << " with val " << *bestVal
            << " on dim " << *bestDim << endl;
     }
   } // newbest
 }
 
-/** Calculate error reduction for this possible split. */
-float LinearSplitsTree::calcER(int dim, float val,
-                               const std::vector<tree_experience*> &instances,
-                               float avgError,
-                               std::vector<tree_experience*> &left,
-                               std::vector<tree_experience*> &right,
-                               float *leftError, float *rightError){
-  if (DTDEBUG || SPLITDEBUG) cout << "calcER, dim=" << dim
+float M5Tree::calcSDR(int dim, float val, 
+                      const std::vector<tree_experience*> &instances,
+                      float sd,
+                      std::vector<tree_experience*> &left,
+                      std::vector<tree_experience*> &right){
+  if (DTDEBUG) cout << "calcSDR, dim=" << dim
                     << " val=" << val
-                    << " err=" << avgError
+                    << " sd=" << sd
                     << " nInstances= " << instances.size() << endl;
 
   left.clear();
@@ -1287,7 +1368,8 @@ float LinearSplitsTree::calcER(int dim, float val,
 
   // split into two sides
   for (unsigned i = 0; i < instances.size(); i++){
-    if (DTDEBUG) cout << " calcER - Classify instance " << i << " on new split " << endl;
+    if (DTDEBUG) cout << "calcSDR - Classify instance " << i 
+                      << " on new split " << endl;
 
     if (passTest(dim, val, instances[i]->input)){
       left.push_back(instances[i]);
@@ -1301,53 +1383,53 @@ float LinearSplitsTree::calcER(int dim, float val,
                     << ", right has " << right.size() << endl;
 
   // get sd for both sides
-  *leftError = calcAvgErrorforSet(left);
-  *rightError = calcAvgErrorforSet(right);
+  float sdLeft = calcSDforSet(left);
+  float sdRight = calcSDforSet(right);
 
   float leftRatio = (float)left.size() / (float)instances.size();
   float rightRatio = (float)right.size() / (float)instances.size();
-  float newError = (leftRatio * (*leftError) + rightRatio * (*rightError));
 
-  float er = avgError - newError;
+  float sdr = sd - (leftRatio * sdLeft + rightRatio * sdRight);
 
-  if (DTDEBUG || SPLITDEBUG){
-    cout << "LeftError: " << *leftError
-         << " RightError: " << *rightError
-         << " NewError: " << newError
-         << " NodeError: " << avgError
-         << " ER: " << er
+  if (DTDEBUG){
+    cout << "LeftSD: " << sdLeft
+         << " RightSD: " << sdRight
+         << " SD: " << sd
+         << " SDR: " << sdr
          << endl;
   }
 
-  return er;
+  return sdr;
 
 }
 
-/** Calculate std deviation for set. */
-float LinearSplitsTree::calcAvgErrorforSet(const std::vector<tree_experience*> &instances){
-  if (DTDEBUG) cout << "calcAvgErrorforSet" << endl;
+float M5Tree::calcSDforSet(const std::vector<tree_experience*> &instances){
+  if (DTDEBUG) cout << "calcSDforSet" << endl;
 
   int n = instances.size();
 
   if (n == 0)
     return 0;
 
-  // fit a linear model to instances
-  // and figure out avg error of it
-  float constant;
-  float avgError = 0.0;
-  std::vector<float> coeff;
-  if (SIMPLE) 
-    avgError = fitSimpleLinearModel(instances, &constant, &coeff);
-  else 
-    avgError = fitMultiLinearModel(instances, &constant, &coeff);
+  double sum = 0;
+  double sumSqr = 0;
 
-  return avgError;
+  // go through instances and calculate sums, sum of squares
+  for (unsigned i = 0; i < instances.size(); i++){
+    float val = instances[i]->output;
+    sum += val;
+    sumSqr += (val * val);
+  }
+
+  double mean = sum / (double)n;
+  double variance = (sumSqr - sum*mean)/(double)n;
+  float sd = sqrt(variance);
+
+  return sd;
+
 }
 
-
-/** Returns the unique elements at this index */
-std::set<float> LinearSplitsTree::getUniques(int dim, const std::vector<tree_experience*> &instances, float& minVal, float& maxVal){
+std::set<float> M5Tree::getUniques(int dim, const std::vector<tree_experience*> &instances, float& minVal, float& maxVal){
   if (DTDEBUG) cout << "getUniques,dim = " << dim;
 
   std::set<float> uniques;
@@ -1361,7 +1443,6 @@ std::set<float> LinearSplitsTree::getUniques(int dim, const std::vector<tree_exp
     uniques.insert(instances[i]->input[dim]);
   }
 
-  
   // lets not try more than 100 possible splits per dimension
   if (uniques.size() > 100){
     float rangeInc = (maxVal - minVal) / 100.0;
@@ -1371,16 +1452,13 @@ std::set<float> LinearSplitsTree::getUniques(int dim, const std::vector<tree_exp
     }
   }
   uniques.insert(maxVal);
-  
 
   if (DTDEBUG) cout << " #: " << uniques.size() << endl;
   return uniques;
 }
 
 
-/** Returns a list of the attributes in this dimension sorted
-    from lowest to highest. */
-float* LinearSplitsTree::sortOnDim(int dim, const std::vector<tree_experience*> &instances){
+float* M5Tree::sortOnDim(int dim, const std::vector<tree_experience*> &instances){
   if (DTDEBUG) cout << "sortOnDim,dim = " << dim << endl;
 
   float* values = new float[instances.size()];
@@ -1435,8 +1513,7 @@ float* LinearSplitsTree::sortOnDim(int dim, const std::vector<tree_experience*> 
 }
 
 
-/** Print the tree for debug purposes. */
-void LinearSplitsTree::printTree(tree_node *t, int level){
+void M5Tree::printTree(tree_node *t, int level){
 
   for (int i = 0; i < level; i++){
     cout << ".";
@@ -1473,8 +1550,7 @@ void LinearSplitsTree::printTree(tree_node *t, int level){
 
 
 // output a map of outcomes and their probabilities for this leaf node
-void LinearSplitsTree::leafPrediction(tree_node* leaf, const std::vector<float> &input,
-                                      std::map<float, float>* retval){
+void M5Tree::leafPrediction(tree_node* leaf, const std::vector<float> &input, std::map<float, float>* retval){
   if (DTDEBUG)
     cout << "Calculating output for leaf " << leaf->id << endl;
 
@@ -1493,13 +1569,12 @@ void LinearSplitsTree::leafPrediction(tree_node* leaf, const std::vector<float> 
   if (DTDEBUG) cout << " prediction: " << prediction << endl;
 
   (*retval)[prediction] = 1.0;
- 
 }
 
 
-void LinearSplitsTree::initNodes(){
+void M5Tree::initNodes(){
 
-  for (int i = 0; i < N_LS_NODES; i++){
+  for (int i = 0; i < N_M5_NODES; i++){
     initTreeNode(&(allNodes[i]));
     freeNodes.push_back(i);
     if (NODEDEBUG) 
@@ -1509,12 +1584,12 @@ void LinearSplitsTree::initNodes(){
 
 }
 
-LinearSplitsTree::tree_node* LinearSplitsTree::allocateNode(){
+M5Tree::tree_node* M5Tree::allocateNode(){
   if (freeNodes.empty()){
     tree_node* newNode = new tree_node;
     initTreeNode(newNode);
     if (NODEDEBUG) 
-      cout << id << " PROBLEM: No more pre-allocated nodes!!!" << endl
+      cout << "PROBLEM: No more pre-allocated nodes!!!" << endl
            << "return new node " << newNode->id 
            << ", now " << freeNodes.size() << " free nodes." << endl;
     return newNode;
@@ -1523,15 +1598,15 @@ LinearSplitsTree::tree_node* LinearSplitsTree::allocateNode(){
   int i = freeNodes.back();
   freeNodes.pop_back();
   if (NODEDEBUG) 
-    cout << id << " allocate node " << i << " with id " << allNodes[i].id 
+    cout << "allocate node " << i << " with id " << allNodes[i].id 
          << ", now " << freeNodes.size() << " free nodes." << endl;
   return &(allNodes[i]);
 }
 
-void LinearSplitsTree::deallocateNode(tree_node* node){
-  if (node->id >= N_LS_NODES){
+void M5Tree::deallocateNode(tree_node* node){
+  if (node->id >= N_M5_NODES){
     if (NODEDEBUG) 
-      cout << id << " dealloc extra node id " << node->id 
+      cout << "dealloc extra node id " << node->id 
            << ", now " << freeNodes.size() << " free nodes." << endl;
     delete node;
     return;
@@ -1539,6 +1614,6 @@ void LinearSplitsTree::deallocateNode(tree_node* node){
 
   freeNodes.push_back(node->id);
   if (NODEDEBUG) 
-    cout << id << " dealloc node " << node->id 
+    cout << "dealloc node " << node->id 
          << ", now " << freeNodes.size() << " free nodes." << endl;
 }
