+++ src/models/Stump.cc
@@ -1,4 +1,4 @@
-#include "Stump.hh"
+#include "models/Stump.hh"
